amend.bayesmanecfit

Modifies an existing bayesmanecfit object, for example, by adding or removing fitted models.

# S3 method for bayesmanecfit
amend(
  object,
  drop,
  add,
  loo_controls,
  x_range = NA,
  precision = 1000,
  sig_val = 0.01,
  priors
)

Arguments

object
An object of class bayesmanecfit, as returned by bnec.

drop
A character vector containing the names of model types you which to drop for the modified fit.

add
A character vector containing the names of model types to add to the modified fit.

loo_controls
A named list of two elements ("fitting" and/or "weights"), each being a named list containing the desired arguments to be passed on to loo (via "fitting") or to loo_model_weights (via "weights"). If "fitting" is provided with argument pointwise = TRUE (due to memory issues) and family = "beta_binomial2", the bnec will fail because that is a custom family. If "weights" is not provided by the user, bnec will set the default method argument in loo_model_weights to "pseudobma". See ?loo_model_weights for further info.

x_range
A range of predictor values over which to consider extracting ECx.

precision
The length of the predictor vector used for posterior predictions, and over which to extract ECx values. Large values will be slower but more precise.

sig_val
Probability value to use as the lower quantile to test significance of the predicted posterior values against the lowest observed concentration (assumed to be the control), to estimate NEC as an interpolated NOEC value from smooth ECx curves.

priors
An object of class brmsprior which specifies user-desired prior distributions of model parameters. If missing, amend will figure out a baseline prior for each parameter. It can also be specified as a named list where each name needs to correspond to the same string as model. See details.

Value

All successfully fitted bayesmanecfit model fits.

Examples

library(bayesnec)
data(manec_example)
exmp <- amend(manec_example, drop = "nec4param")
#> Only ecx4param is fitted, no model averaging done. Perhaps try setting better priors, or check ?show_params to make sure you have the correct parameter names for your priors.
#> Returning ecx4param
#> Warning: Found 2 observations with a pareto_k > 0.7 in model 'ecx4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 5 (5.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.