Class bayesmanecfit of models fitted with the brms package

Multiple models fitted with the bayesnec package are represented as a bayesmanecfit object, which contains the original brmsfit fitted objects, names of non-linear models that were fitted, model averaging WAIC stats, sample size, mean posterior NEC values, mean model averaged predictions on the data scale, model averaged residuals, full posterior distribution of predicated values, and summary statistics of NEC statistics.

Details

See methods(class = "bayesmanecfit") for an overview of available methods.

Slots

mod_fits

A list of fitted model outputs of class prebayesnecfit for each of the fitted models.

success_models

A character vector indicating the name of the successfully fitted models.

mod_stats

A data.frame of model fit statistics.

sample_size

The size of the posterior sample. information on the priors used in the model.

w_nec_posterior

The model-weighted posterior estimate of the NEC.

w_predicted_y

The model-weighted predicted values for the observed data.

w_residuals

Model-weighted residual values (i.e. observed - w_predicted_y).

w_pred_vals

A list containing model-weighted posterior predicted values based on the supplied precision and x_range.

w_nec

The summary stats (median and 95% credibility intervals) of w_nec_posterior.

See also