Class bayesnecfit of models fitted with the brms package

Models fitted with the bayesnec package are represented as a bayesnecfit object, which contain the original brmsfit fitted object, list of initialisation values used, the validated bayesnecformula, name of non-linear model that was fitted, posterior predictions, posterior parameter estimates and a series of other statistics.

Details

See methods(class = "bayesnecfit") for an overview of available methods.

Slots

fit

The fitted Bayesian model of class brmsfit.

model

A character string indicating the name of the fitted model.

inits

A list containing the initialisation values for to fit the model.

bayesnecformula

An object of class bayesnecformula and formula.

pred_vals

A list containing a data.frame of summary posterior predicted values and a vector containing based on the supplied precision and x_range.

top

The estimate for parameter "top" in the fitted model.

beta

The estimate for parameter "beta" in the fitted model.

nec

The estimated NEC.

f

The estimate for parameter "f" in the fitted model, NA if absent for the fitted model type.

bot

The estimate for parameter "bot" in the fitted model, NA if absent for the fitted model type.

d

The estimate for parameter "d" in the fitted model, NA if absent for the fitted model type.

slope

The estimate for parameter "slope" in the fitted model, NA if absent for the fitted model type.

ec50

The estimate for parameter "ec50" in the fitted model, NA if absent for the fitted model type.

dispersion

An estimate of dispersion.

predicted_y

The predicted values for the observed data.

residuals

Residual values of the observed data from the fitted model.

nec_posterior

A full posterior estimate of the NEC.

See also