check_data

Check data input for a Bayesian NEC model fit

check_data(data, x_var, y_var, trials_var, x_type = NA, family = NA, model)

Arguments

data
A data.frame containing the data to use for the model.

x_var
A character indicating the column heading containing the concentration (x) variable.

y_var
A character indicating the column heading containing the response (y) variable.

trials_var
The column heading indicating the column for the number of "trials" for binomial response data. If not supplied, the model may run but will not be the model you intended!

x_type
The statistical distribution of the x (concentration) data. This will be guessed based on the characteristic of the input data if not supplied and will inform the priors used for estimating nec in NEC step models.

family
Either a character string, a function, or an object of class family defining the statistical distribution (family) to use for the y (response) data. See details.

model
A character vector indicating the model(s) to fit. See Details for more information.

Value

A list of modified elements necessary for fit_bayesnec.

Details

This is a wrapper function to test input data criteria and find the correct priors for use in fit_bayesnec.