compare_endpoints

Extracts posterior predicted values from a list of class bayesnecfit or bayesnecfit model fits and compares these via bootstrap re sampling.

compare_endpoints(
  x,
  comparison = "nec",
  ecx_val = 10,
  type = "absolute",
  hormesis_def = "control",
  sig_val = 0.01,
  precision,
  x_range = NA
)

Arguments

x
A named list of objects of class bayesnecfit or bayesnecfit returned by bnec.

comparison
The posterior predictions to compare, takes values of "nec", "nsec", "ecx" or "fitted".

ecx_val
The desired percentage effect value. This must be a value between 1 and 99 (for type = "relative" and "absolute"), defaults to 10.

type
A character vector, taking values of "relative", "absolute" (the default) or "direct". See Details.

hormesis_def
A character vector, taking values of "max" or "control". See Details.

sig_val
Probability value to use as the lower quantile to test significance of the predicted posterior values. against the lowest observed concentration (assumed to be the control), to estimate NEC as an interpolated NOEC value from smooth ECx curves.

precision
The number of unique x values over which to find ECx - large values will make the ECx estimate more precise.

x_range
A range of x values over which to consider extracting ECx.

Value

A named list containing bootstrapped differences in posterior predictions of the bayesnecfit or bayesnecfit model fits contained in x. See Details.

See also

Examples

library(bayesnec)
data(manec_example)
ecx4param <- pull_out(manec_example, model = "ecx4param")
#> Pulling out model(s): ecx4param
nec4param <- pull_out(manec_example, model = "nec4param")
#> Pulling out model(s): nec4param
compare_endpoints(list("nec" = ecx4param, "ecx" = nec4param), ecx_val = 50)
#> $posterior_list
#> $posterior_list$nec
#>   [1] 0.77912696 0.80784923 0.73763924 0.81104059 0.85571967 0.78870105
#>   [7] 0.42169430 0.60998472 0.91635557 0.65466380 0.98656555 0.76636151
#>  [13] 0.94188647 0.79189241 0.72487379 0.93550375 1.10145462 0.89401603
#>  [19] 0.75997878 0.75997878 0.59083654 0.59721927 0.98656555 1.01528782
#>  [25] 0.67062062 0.84614558 0.97380010 0.92273829 0.71849106 0.89082466
#>  [31] 0.94188647 0.90678148 0.79827514 0.92912102 0.76317014 0.71849106
#>  [37] 0.81423195 0.86210240 0.76955287 0.83976286 0.77912696 0.97699147
#>  [43] 0.98975692 0.84614558 1.01209646 0.75040469 0.82699740 1.02167055
#>  [49] 0.88763330 0.98018283 0.98975692 1.01847918 0.89720739 0.87486785
#>  [55] 0.93869511 0.68019470 1.01209646 0.67062062 0.87486785 0.78550968
#>  [61] 0.77912696 0.88444194 0.78231832 0.97380010 0.89082466 0.76317014
#>  [67] 0.93231238 0.69296016 0.90997284 0.80784923 0.69934288 0.91316420
#>  [73] 0.91316420 0.92273829 0.81104059 0.83018877 0.92912102 0.92912102
#>  [79] 0.91316420 0.92273829 0.75678742 0.85252831 1.04720145 1.00890510
#>  [85] 0.87167649 0.03234801 0.51743519 0.62913290 0.85252831 0.77593560
#>  [91] 0.83976286 0.98018283 0.89082466 0.86529376 0.03873074 0.94507783
#>  [97] 0.79189241 0.86848512 0.56849700 0.96422601
#> 
#> $posterior_list$ecx
#>   [1] 1.442587 1.457722 1.481493 1.485925 1.458289 1.449933 1.443140 1.466578
#>   [9] 1.436326 1.466663 1.421838 1.514186 1.429119 1.440516 1.397421 1.509695
#>  [17] 1.523597 1.530270 1.466546 1.466265 1.459125 1.431784 1.508234 1.508211
#>  [25] 1.419997 1.455781 1.432226 1.443483 1.435387 1.438993 1.405659 1.418535
#>  [33] 1.412248 1.417279 1.425567 1.436740 1.441224 1.415182 1.452339 1.420860
#>  [41] 1.423742 1.447464 1.439129 1.470685 1.447541 1.504913 1.524641 1.510200
#>  [49] 1.521482 1.522349 1.484843 1.396919 1.477415 1.473241 1.431799 1.421336
#>  [57] 1.434694 1.442501 1.425624 1.453501 1.475282 1.421443 1.469332 1.501495
#>  [65] 1.472506 1.350501 1.495111 1.394531 1.439682 1.487531 1.454148 1.429012
#>  [73] 1.508282 1.465438 1.433031 1.497852 1.486412 1.473494 1.474565 1.440962
#>  [81] 1.416739 1.494570 1.477715 1.364731 1.417989 1.495071 1.407532 1.489753
#>  [89] 1.465615 1.492940 1.472939 1.469240 1.508738 1.433194 1.407564 1.472851
#>  [97] 1.405898 1.421342 1.428190 1.469698
#> 
#> 
#> $posterior_data
#>     model      value
#> 1     ecx 1.45414848
#> 2     ecx 1.47456464
#> 3     ecx 1.41224849
#> 4     ecx 1.44250084
#> 5     ecx 1.44348314
#> 6     ecx 1.43303115
#> 7     ecx 1.43178401
#> 8     ecx 1.45828884
#> 9     ecx 1.36473055
#> 10    ecx 1.50828232
#> 11    ecx 1.42556730
#> 12    ecx 1.39691939
#> 13    ecx 1.40753180
#> 14    ecx 1.48149289
#> 15    ecx 1.46654550
#> 16    ecx 1.48592514
#> 17    ecx 1.48484278
#> 18    ecx 1.47324054
#> 19    ecx 1.48975297
#> 20    ecx 1.43899257
#> 21    ecx 1.49507088
#> 22    ecx 1.44754067
#> 23    ecx 1.43968246
#> 24    ecx 1.41518219
#> 25    ecx 1.44993293
#> 26    ecx 1.50969467
#> 27    ecx 1.44314048
#> 28    ecx 1.42086034
#> 29    ecx 1.46543824
#> 30    ecx 1.40756436
#> 31    ecx 1.49294026
#> 32    ecx 1.42819014
#> 33    ecx 1.39453145
#> 34    ecx 1.47528161
#> 35    ecx 1.43222552
#> 36    ecx 1.45578075
#> 37    ecx 1.42144280
#> 38    ecx 1.45233871
#> 39    ecx 1.52359671
#> 40    ecx 1.42374232
#> 41    ecx 1.47771545
#> 42    ecx 1.47349375
#> 43    ecx 1.47068468
#> 44    ecx 1.49457037
#> 45    ecx 1.51418600
#> 46    ecx 1.45912481
#> 47    ecx 1.44096171
#> 48    ecx 1.46561543
#> 49    ecx 1.42183784
#> 50    ecx 1.43179896
#> 51    ecx 1.42133560
#> 52    ecx 1.47741463
#> 53    ecx 1.50823392
#> 54    ecx 1.43912868
#> 55    ecx 1.39742123
#> 56    ecx 1.41999697
#> 57    ecx 1.46969803
#> 58    ecx 1.50491309
#> 59    ecx 1.50873821
#> 60    ecx 1.45772230
#> 61    ecx 1.42562379
#> 62    ecx 1.43538704
#> 63    ecx 1.53026950
#> 64    ecx 1.46666288
#> 65    ecx 1.52148153
#> 66    ecx 1.44051576
#> 67    ecx 1.51020005
#> 68    ecx 1.42134157
#> 69    ecx 1.41798904
#> 70    ecx 1.48641153
#> 71    ecx 1.43469447
#> 72    ecx 1.43674021
#> 73    ecx 1.46933243
#> 74    ecx 1.42901223
#> 75    ecx 1.49785177
#> 76    ecx 1.35050091
#> 77    ecx 1.43632564
#> 78    ecx 1.41673923
#> 79    ecx 1.52464099
#> 80    ecx 1.40589762
#> 81    ecx 1.41853495
#> 82    ecx 1.47293908
#> 83    ecx 1.46657821
#> 84    ecx 1.42911946
#> 85    ecx 1.44746431
#> 86    ecx 1.52234899
#> 87    ecx 1.46626497
#> 88    ecx 1.45350109
#> 89    ecx 1.40565929
#> 90    ecx 1.43319441
#> 91    ecx 1.41727890
#> 92    ecx 1.47285106
#> 93    ecx 1.50149535
#> 94    ecx 1.44258670
#> 95    ecx 1.50821071
#> 96    ecx 1.47250632
#> 97    ecx 1.46923980
#> 98    ecx 1.48753061
#> 99    ecx 1.49511064
#> 100   ecx 1.44122407
#> 101   nec 0.87167649
#> 102   nec 0.81104059
#> 103   nec 0.89082466
#> 104   nec 0.78550968
#> 105   nec 0.69934288
#> 106   nec 0.85252831
#> 107   nec 0.85252831
#> 108   nec 0.89401603
#> 109   nec 0.80784923
#> 110   nec 0.98018283
#> 111   nec 0.76955287
#> 112   nec 0.76317014
#> 113   nec 0.75997878
#> 114   nec 0.78231832
#> 115   nec 0.93231238
#> 116   nec 0.83976286
#> 117   nec 1.01209646
#> 118   nec 0.76317014
#> 119   nec 0.97380010
#> 120   nec 0.89720739
#> 121   nec 0.88444194
#> 122   nec 0.90997284
#> 123   nec 0.79827514
#> 124   nec 1.01209646
#> 125   nec 0.67062062
#> 126   nec 0.91635557
#> 127   nec 0.84614558
#> 128   nec 0.42169430
#> 129   nec 0.85571967
#> 130   nec 0.56849700
#> 131   nec 0.68019470
#> 132   nec 0.92912102
#> 133   nec 0.92273829
#> 134   nec 0.80784923
#> 135   nec 0.84614558
#> 136   nec 0.81104059
#> 137   nec 0.83976286
#> 138   nec 0.94188647
#> 139   nec 0.97699147
#> 140   nec 1.00890510
#> 141   nec 0.96422601
#> 142   nec 0.94188647
#> 143   nec 0.71849106
#> 144   nec 0.77912696
#> 145   nec 1.02167055
#> 146   nec 0.59721927
#> 147   nec 1.10145462
#> 148   nec 0.79189241
#> 149   nec 0.65466380
#> 150   nec 0.94507783
#> 151   nec 1.04720145
#> 152   nec 0.69296016
#> 153   nec 0.89082466
#> 154   nec 0.89082466
#> 155   nec 0.92273829
#> 156   nec 0.97380010
#> 157   nec 0.86848512
#> 158   nec 0.60998472
#> 159   nec 0.03873074
#> 160   nec 0.81423195
#> 161   nec 0.98975692
#> 162   nec 0.59083654
#> 163   nec 0.76636151
#> 164   nec 0.92273829
#> 165   nec 0.87486785
#> 166   nec 0.75678742
#> 167   nec 0.87486785
#> 168   nec 0.98656555
#> 169   nec 0.90678148
#> 170   nec 0.71849106
#> 171   nec 0.98018283
#> 172   nec 0.77912696
#> 173   nec 0.92912102
#> 174   nec 0.77593560
#> 175   nec 0.79189241
#> 176   nec 0.67062062
#> 177   nec 1.01847918
#> 178   nec 0.91316420
#> 179   nec 0.92912102
#> 180   nec 0.77912696
#> 181   nec 0.88763330
#> 182   nec 0.03234801
#> 183   nec 1.01528782
#> 184   nec 0.98975692
#> 185   nec 0.75040469
#> 186   nec 0.86210240
#> 187   nec 0.93869511
#> 188   nec 0.78870105
#> 189   nec 0.98656555
#> 190   nec 0.91316420
#> 191   nec 0.86529376
#> 192   nec 0.91316420
#> 193   nec 0.82699740
#> 194   nec 0.62913290
#> 195   nec 0.93550375
#> 196   nec 0.73763924
#> 197   nec 0.72487379
#> 198   nec 0.75997878
#> 199   nec 0.83018877
#> 200   nec 0.51743519
#> 
#> $diff_list
#> $diff_list$`nec-ecx`
#>   [1] -0.5824720 -0.6635240 -0.5214238 -0.6569912 -0.7441403 -0.5805028
#>   [7] -0.5792557 -0.5642728 -0.5568813 -0.5280995 -0.6560144 -0.6337492
#>  [13] -0.6475530 -0.6991746 -0.5342331 -0.6461623 -0.4727463 -0.7100704
#>  [19] -0.5159529 -0.5417852 -0.6106289 -0.5375678 -0.6414073 -0.4030857
#>  [25] -0.7793123 -0.5933391 -0.5969949 -0.9991660 -0.6097186 -0.8390674
#>  [31] -0.8127456 -0.4990691 -0.4717932 -0.6674324 -0.5860799 -0.6447402
#>  [37] -0.5816799 -0.5104522 -0.5466052 -0.4148372 -0.5134894 -0.5316073
#>  [43] -0.7521936 -0.7154434 -0.4925155 -0.8619055 -0.3395071 -0.6737230
#>  [49] -0.7671740 -0.4867211 -0.3741342 -0.7844545 -0.6174093 -0.5483040
#>  [55] -0.4746829 -0.4461969 -0.6012129 -0.8949284 -1.4700075 -0.6434904
#>  [61] -0.4358669 -0.8445505 -0.7639080 -0.5439246 -0.6466137 -0.6837283
#>  [67] -0.6353322 -0.4347760 -0.5112076 -0.7679205 -0.4545116 -0.6576133
#>  [73] -0.5402114 -0.6530766 -0.7059594 -0.6798803 -0.4178465 -0.5035750
#>  [79] -0.5955200 -0.6267707 -0.5309017 -1.4405911 -0.4512904 -0.4393625
#>  [85] -0.6970596 -0.6602466 -0.5275699 -0.6648000 -0.4190937 -0.5200302
#>  [91] -0.5519851 -0.5596869 -0.6744979 -0.8134538 -0.5727070 -0.7348671
#>  [97] -0.7443660 -0.7275518 -0.6649219 -0.9237889
#> 
#> 
#> $diff_data
#>     comparison       diff
#> 1      nec-ecx -0.5824720
#> 2      nec-ecx -0.6635240
#> 3      nec-ecx -0.5214238
#> 4      nec-ecx -0.6569912
#> 5      nec-ecx -0.7441403
#> 6      nec-ecx -0.5805028
#> 7      nec-ecx -0.5792557
#> 8      nec-ecx -0.5642728
#> 9      nec-ecx -0.5568813
#> 10     nec-ecx -0.5280995
#> 11     nec-ecx -0.6560144
#> 12     nec-ecx -0.6337492
#> 13     nec-ecx -0.6475530
#> 14     nec-ecx -0.6991746
#> 15     nec-ecx -0.5342331
#> 16     nec-ecx -0.6461623
#> 17     nec-ecx -0.4727463
#> 18     nec-ecx -0.7100704
#> 19     nec-ecx -0.5159529
#> 20     nec-ecx -0.5417852
#> 21     nec-ecx -0.6106289
#> 22     nec-ecx -0.5375678
#> 23     nec-ecx -0.6414073
#> 24     nec-ecx -0.4030857
#> 25     nec-ecx -0.7793123
#> 26     nec-ecx -0.5933391
#> 27     nec-ecx -0.5969949
#> 28     nec-ecx -0.9991660
#> 29     nec-ecx -0.6097186
#> 30     nec-ecx -0.8390674
#> 31     nec-ecx -0.8127456
#> 32     nec-ecx -0.4990691
#> 33     nec-ecx -0.4717932
#> 34     nec-ecx -0.6674324
#> 35     nec-ecx -0.5860799
#> 36     nec-ecx -0.6447402
#> 37     nec-ecx -0.5816799
#> 38     nec-ecx -0.5104522
#> 39     nec-ecx -0.5466052
#> 40     nec-ecx -0.4148372
#> 41     nec-ecx -0.5134894
#> 42     nec-ecx -0.5316073
#> 43     nec-ecx -0.7521936
#> 44     nec-ecx -0.7154434
#> 45     nec-ecx -0.4925155
#> 46     nec-ecx -0.8619055
#> 47     nec-ecx -0.3395071
#> 48     nec-ecx -0.6737230
#> 49     nec-ecx -0.7671740
#> 50     nec-ecx -0.4867211
#> 51     nec-ecx -0.3741342
#> 52     nec-ecx -0.7844545
#> 53     nec-ecx -0.6174093
#> 54     nec-ecx -0.5483040
#> 55     nec-ecx -0.4746829
#> 56     nec-ecx -0.4461969
#> 57     nec-ecx -0.6012129
#> 58     nec-ecx -0.8949284
#> 59     nec-ecx -1.4700075
#> 60     nec-ecx -0.6434904
#> 61     nec-ecx -0.4358669
#> 62     nec-ecx -0.8445505
#> 63     nec-ecx -0.7639080
#> 64     nec-ecx -0.5439246
#> 65     nec-ecx -0.6466137
#> 66     nec-ecx -0.6837283
#> 67     nec-ecx -0.6353322
#> 68     nec-ecx -0.4347760
#> 69     nec-ecx -0.5112076
#> 70     nec-ecx -0.7679205
#> 71     nec-ecx -0.4545116
#> 72     nec-ecx -0.6576133
#> 73     nec-ecx -0.5402114
#> 74     nec-ecx -0.6530766
#> 75     nec-ecx -0.7059594
#> 76     nec-ecx -0.6798803
#> 77     nec-ecx -0.4178465
#> 78     nec-ecx -0.5035750
#> 79     nec-ecx -0.5955200
#> 80     nec-ecx -0.6267707
#> 81     nec-ecx -0.5309017
#> 82     nec-ecx -1.4405911
#> 83     nec-ecx -0.4512904
#> 84     nec-ecx -0.4393625
#> 85     nec-ecx -0.6970596
#> 86     nec-ecx -0.6602466
#> 87     nec-ecx -0.5275699
#> 88     nec-ecx -0.6648000
#> 89     nec-ecx -0.4190937
#> 90     nec-ecx -0.5200302
#> 91     nec-ecx -0.5519851
#> 92     nec-ecx -0.5596869
#> 93     nec-ecx -0.6744979
#> 94     nec-ecx -0.8134538
#> 95     nec-ecx -0.5727070
#> 96     nec-ecx -0.7348671
#> 97     nec-ecx -0.7443660
#> 98     nec-ecx -0.7275518
#> 99     nec-ecx -0.6649219
#> 100    nec-ecx -0.9237889
#> 
#> $prob_diff
#>   comparison prob
#> 1    nec-ecx    0
#>