compare_fitted

Extracts posterior predicted values from a list of class bayesnecfit or bayesnecfit model fits and compares these across a vector of fitted values.

compare_fitted(x, precision = 50, x_range = NA)

Arguments

x
A named list list of objects of class bayesnecfit or bayesnecfit returned by bnec.

precision
The number of unique x values over which to find ECx - large values will make the ECx estimate more precise.

x_range
A range of x values over which to consider extracting ECx.

Value

A named list containing bootstrapped differences in posterior predictions of the bayesnecfit or bayesnecfit model fits contained in x. See Details.

See also

Examples

# \donttest{
library(bayesnec)
data(manec_example)
ecx4param <- pull_out(manec_example, model = "ecx4param")
#> Pulling out model(s): ecx4param
nec4param <- pull_out(manec_example, model = "nec4param")
#> Pulling out model(s): nec4param
compare_fitted(list("nec" = ecx4param, "ecx" = nec4param))
#> $posterior_list
#> $posterior_list$nec
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 2.330578 2.325562 2.319689 2.312815 2.304772 2.295365 2.284366 2.271511
#>  [2,] 2.272395 2.266392 2.259464 2.251471 2.242253 2.231625 2.219375 2.205265
#>  [3,] 2.257957 2.256619 2.254971 2.252942 2.250445 2.247371 2.243589 2.238935
#>  [4,] 2.204103 2.203072 2.201787 2.200186 2.198191 2.195707 2.192612 2.188758
#>  [5,] 2.177634 2.176018 2.174045 2.171636 2.168697 2.165109 2.160732 2.155393
#>  [6,] 2.276866 2.274554 2.271755 2.268365 2.264262 2.259296 2.253287 2.246019
#>  [7,] 2.284438 2.280882 2.276661 2.271651 2.265707 2.258655 2.250293 2.240381
#>  [8,] 2.314643 2.310027 2.304600 2.298219 2.290721 2.281913 2.271569 2.259429
#>  [9,] 2.172208 2.170379 2.168159 2.165466 2.162200 2.158240 2.153438 2.147618
#> [10,] 2.323276 2.321585 2.319513 2.316975 2.313867 2.310062 2.305404 2.299702
#>           [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>  [1,] 2.256497 2.238969 2.218524 2.194695 2.166951 2.134684 2.097208 2.053750
#>  [2,] 2.189019 2.170325 2.148831 2.124137 2.095793 2.063295 2.026080 1.983524
#>  [3,] 2.233211 2.226173 2.217522 2.206894 2.193846 2.177839 2.158217 2.134194
#>  [4,] 2.183960 2.177989 2.170560 2.161320 2.149837 2.135573 2.117872 2.095930
#>  [5,] 2.148882 2.140945 2.131274 2.119496 2.105162 2.087731 2.066554 2.040856
#>  [6,] 2.237232 2.226613 2.213789 2.198311 2.179649 2.157170 2.130127 2.097644
#>  [7,] 2.228639 2.214735 2.198284 2.178835 2.155864 2.128764 2.096835 2.059279
#>  [8,] 2.245188 2.228493 2.208939 2.186054 2.159302 2.128067 2.091653 2.049272
#>  [9,] 2.140566 2.132024 2.121683 2.109170 2.094041 2.075764 2.053705 2.027115
#> [10,] 2.292727 2.284196 2.273768 2.261028 2.245477 2.226508 2.203398 2.175278
#>          [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#>  [1,] 2.003446 1.945338 1.878378 1.801432 1.713292 1.612698 1.498368 1.369042
#>  [2,] 1.934938 1.879570 1.816607 1.745178 1.664363 1.573209 1.470750 1.356030
#>  [3,] 2.104821 2.068967 2.025292 1.972221 1.907928 1.830324 1.737065 1.625586
#>  [4,] 2.068764 2.035188 1.993771 1.942812 1.880304 1.803917 1.710999 1.598604
#>  [5,] 2.009717 1.972048 1.926575 1.871820 1.806086 1.727458 1.633813 1.522858
#>  [6,] 2.058699 2.012107 1.956514 1.890388 1.812027 1.719575 1.611066 1.484485
#>  [7,] 2.015185 1.963528 1.903168 1.832850 1.751217 1.656833 1.548219 1.423902
#>  [8,] 2.000044 1.942996 1.877061 1.801089 1.713865 1.614129 1.500618 1.372113
#>  [9,] 1.995111 1.956657 1.910554 1.855419 1.789686 1.711601 1.619242 1.510559
#> [10,] 2.141118 2.099702 2.049605 1.989182 1.916556 1.829621 1.726074 1.603465
#>          [,25]    [,26]     [,27]     [,28]     [,29]     [,30]        [,31]
#>  [1,] 1.223536 1.060808 0.8800432 0.6807405 0.4628042 0.2266317 -0.026814667
#>  [2,] 1.228138 1.086255 0.9296922 0.7579547 0.5707934 0.3682648  0.150781905
#>  [3,] 1.493169 1.337059 1.1546395 0.9436699 0.7025805 0.4308053  0.129111423
#>  [4,] 1.463564 1.302631 1.1126861 0.8910414 0.6358233 0.3464077  0.023850591
#>  [5,] 1.392195 1.239426 1.0623076 0.8589490 0.6280583 0.3692119  0.083118355
#>  [6,] 1.337871 1.169449 0.9777960 0.7620434 0.5220841 0.2587699 -0.025939092
#>  [7,] 1.282489 1.122756 0.9437579 0.7449458 0.5262958 0.2884239  0.032675836
#>  [8,] 1.227505 1.065873 0.8865719 0.6893289 0.4743379 0.2423413 -0.005311142
#>  [9,] 1.383432 1.235771 1.0656531 0.8714957 0.6522622 0.4076877  0.138494707
#> [10,] 1.459296 1.291160 1.0969376 0.8750453 0.6247142 0.3462816  0.041442533
#>             [,32]      [,33]      [,34]      [,35]     [,36]     [,37]
#>  [1,] -0.29596268 -0.5786225 -0.8720250 -1.1729087 -1.477653 -1.782445
#>  [2,] -0.08084389 -0.3253784 -0.5811627 -0.8461369 -1.117893 -1.393754
#>  [3,] -0.20013391 -0.5528289 -0.9231993 -1.3040913 -1.687481 -2.065135
#>  [4,] -0.32877346 -0.7062282 -1.1012565 -1.5050427 -1.907955 -2.300444
#>  [5,] -0.22816771 -0.5611353 -0.9108578 -1.2712120 -1.635270 -1.995813
#>  [6,] -0.32891049 -0.6458929 -0.9717026 -1.3005358 -1.626370 -1.943401
#>  [7,] -0.23882678 -0.5232017 -0.8168916 -1.1158208 -1.415611 -1.711833
#>  [8,] -0.26664016 -0.5390618 -0.8194640 -1.1043338 -1.389926 -1.672459
#>  [9,] -0.15343613 -0.4649810 -0.7918104 -1.1285696 -1.469198 -1.807352
#> [10,] -0.28659405 -0.6330991 -0.9920069 -1.3562913 -1.718494 -2.071326
#>           [,38]     [,39]     [,40]     [,41]     [,42]     [,43]     [,44]
#>  [1,] -2.083472 -2.377105 -2.660073 -2.929592 -3.183457 -3.420080 -3.638485
#>  [2,] -1.670875 -1.946358 -2.217374 -2.481276 -2.735698 -2.978629 -3.208465
#>  [3,] -2.429299 -2.773317 -3.092054 -3.382099 -3.641744 -3.870784 -4.070225
#>  [4,] -2.673925 -3.021480 -3.338258 -3.621565 -3.870674 -4.086467 -4.270995
#>  [5,] -2.345899 -2.679377 -2.991286 -3.278076 -3.537655 -3.769282 -3.973353
#>  [6,] -2.246440 -2.531233 -2.794641 -3.034700 -3.250554 -3.442306 -3.610822
#>  [7,] -2.000265 -2.277129 -2.539265 -2.784254 -3.010451 -3.216967 -3.403590
#>  [8,] -1.948308 -2.214192 -2.467315 -2.705460 -2.927040 -3.131085 -3.317203
#>  [9,] -2.136868 -2.452206 -2.748786 -3.023212 -3.273330 -3.498176 -3.697812
#> [10,] -2.408237 -2.723870 -3.014331 -3.277263 -3.511754 -3.718115 -3.897603
#>           [,45]     [,46]     [,47]     [,48]     [,49]     [,50]
#>  [1,] -3.838259 -4.019487 -4.182662 -4.328592 -4.458312 -4.573005
#>  [2,] -3.424029 -3.624563 -3.809702 -3.979434 -4.134044 -4.274054
#>  [3,] -4.241940 -4.388351 -4.512153 -4.616104 -4.702873 -4.774944
#>  [4,] -4.427047 -4.557787 -4.666461 -4.756205 -4.829917 -4.890193
#>  [5,] -4.151141 -4.304524 -4.435739 -4.547182 -4.641253 -4.720249
#>  [6,] -3.757523 -3.884183 -3.992764 -4.085279 -4.163695 -4.229868
#>  [7,] -3.570682 -3.719049 -3.849821 -3.964338 -4.064050 -4.150441
#>  [8,] -3.485499 -3.636487 -3.770996 -3.890072 -3.994900 -4.086734
#>  [9,] -3.873124 -4.025594 -4.157085 -4.269664 -4.365451 -4.446522
#> [10,] -4.052135 -4.184015 -4.295723 -4.389743 -4.468453 -4.534052
#> 
#> $posterior_list$ecx
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603
#>  [2,] 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850
#>  [3,] 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850
#>  [4,] 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147
#>  [5,] 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286
#>  [6,] 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616
#>  [7,] 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557
#>  [8,] 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777
#>  [9,] 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511
#> [10,] 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204
#>           [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>  [1,] 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603
#>  [2,] 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850
#>  [3,] 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850
#>  [4,] 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147
#>  [5,] 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286
#>  [6,] 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616
#>  [7,] 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557
#>  [8,] 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777
#>  [9,] 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511
#> [10,] 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204
#>          [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#>  [1,] 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603 2.194603 1.975858
#>  [2,] 2.122850 2.122850 2.122850 2.122850 2.122850 2.122850 2.005619 1.684963
#>  [3,] 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850 2.203850 1.935972
#>  [4,] 2.129147 2.129147 2.129147 2.129147 2.129147 2.129147 1.967766 1.651810
#>  [5,] 2.216286 2.216286 2.216286 2.216286 2.216286 2.216286 2.153047 1.806073
#>  [6,] 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616 2.117616 1.873404
#>  [7,] 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557 2.058557 1.879968
#>  [8,] 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777 2.107777 1.916964
#>  [9,] 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511 2.200511 1.936556
#> [10,] 2.163204 2.163204 2.163204 2.163204 2.163204 2.163204 2.087849 1.730037
#>          [,25]    [,26]     [,27]     [,28]     [,29]        [,30]      [,31]
#>  [1,] 1.626765 1.288234 0.9599452 0.6415891 0.3328650  0.033481481 -0.2568441
#>  [2,] 1.373125 1.069862 0.7749393 0.4881272 0.2092026 -0.062051500 -0.3258459
#>  [3,] 1.594010 1.262165 0.9401404 0.6276440 0.3243942  0.030117381 -0.2554519
#>  [4,] 1.344374 1.045230 0.7541524 0.4709248 0.1953351 -0.072822563 -0.3337486
#>  [5,] 1.469342 1.142552 0.8254081 0.5176266 0.2189309 -0.070947309 -0.3522682
#>  [6,] 1.522424 1.183647 0.8566462 0.5410139 0.2363545 -0.057713497 -0.3415583
#>  [7,] 1.529211 1.190685 0.8639630 0.5486332 0.2442985 -0.049424344 -0.3329055
#>  [8,] 1.561844 1.219219 0.8886512 0.5697151 0.2620017 -0.034883767 -0.3213224
#>  [9,] 1.578653 1.235168 0.9055209 0.5891535 0.2855312 -0.005859621 -0.2855116
#> [10,] 1.386319 1.056139 0.7389645 0.4342832 0.1416029 -0.139549077 -0.4096268
#>            [,32]      [,33]     [,34]     [,35]     [,36]     [,37]     [,38]
#>  [1,] -0.5383857 -0.8114092 -1.076172 -1.332925 -1.581909 -1.823361 -2.057507
#>  [2,] -0.5823859 -0.8318708 -1.074495 -1.310446 -1.539909 -1.763062 -1.980078
#>  [3,] -0.5325713 -0.8014909 -1.062453 -1.315694 -1.561441 -1.799917 -2.031337
#>  [4,] -0.5876381 -0.8346807 -1.075061 -1.308959 -1.536549 -1.758002 -1.973482
#>  [5,] -0.6252845 -0.8902413 -1.147376 -1.396921 -1.639099 -1.874128 -2.102218
#>  [6,] -0.6155354 -0.8799877 -1.135246 -1.381631 -1.619450 -1.849002 -2.070573
#>  [7,] -0.6065019 -0.8705585 -1.125408 -1.371371 -1.608757 -1.837866 -2.058987
#>  [8,] -0.5976819 -0.8643168 -1.121569 -1.369770 -1.609236 -1.840276 -2.063187
#>  [9,] -0.5538976 -0.8114716 -1.058669 -1.295908 -1.523590 -1.742099 -1.951806
#> [10,] -0.6690666 -0.9182873 -1.157692 -1.387666 -1.608582 -1.820796 -2.024652
#>           [,39]     [,40]     [,41]     [,42]     [,43]     [,44]     [,45]
#>  [1,] -2.284568 -2.504761 -2.718291 -2.925360 -3.126165 -3.320894 -3.509732
#>  [2,] -2.191125 -2.396369 -2.595968 -2.790079 -2.978851 -3.162432 -3.340964
#>  [3,] -2.255909 -2.473835 -2.685314 -2.890535 -3.089684 -3.282939 -3.470477
#>  [4,] -2.183152 -2.387167 -2.585681 -2.778841 -2.966791 -3.149674 -3.327624
#>  [5,] -2.323576 -2.538399 -2.746880 -2.949207 -3.145561 -3.336119 -3.521051
#>  [6,] -2.284442 -2.490875 -2.690132 -2.882461 -3.068105 -3.247294 -3.420254
#>  [7,] -2.272397 -2.478365 -2.677152 -2.869008 -3.054174 -3.232883 -3.405361
#>  [8,] -2.278253 -2.485752 -2.685950 -2.879102 -3.065458 -3.245257 -3.418729
#>  [9,] -2.153064 -2.346215 -2.531585 -2.709486 -2.880221 -3.044078 -3.201334
#> [10,] -2.220478 -2.408591 -2.589294 -2.762879 -2.929627 -3.089807 -3.243678
#>           [,46]     [,47]     [,48]     [,49]     [,50]
#>  [1,] -3.692856 -3.870440 -4.042651 -4.209651 -4.371599
#>  [2,] -3.514587 -3.683435 -3.847639 -4.007328 -4.162625
#>  [3,] -3.652465 -3.829069 -4.000447 -4.166754 -4.328139
#>  [4,] -3.500775 -3.669257 -3.833196 -3.992713 -4.147928
#>  [5,] -3.700525 -3.874700 -4.043733 -4.207777 -4.366978
#>  [6,] -3.587202 -3.748345 -3.903886 -4.054020 -4.198934
#>  [7,] -3.571824 -3.732484 -3.887541 -4.037192 -4.181624
#>  [8,] -3.586096 -3.747575 -3.903371 -4.053685 -4.198710
#>  [9,] -3.352255 -3.497095 -3.636101 -3.769507 -3.897538
#> [10,] -3.391488 -3.533476 -3.669872 -3.800894 -3.926757
#> 
#> 
#> $posterior_data
#>     model          x    Estimate        Q2.5       Q97.5
#> 1     nec 0.03234801  2.27463021  2.17342893  2.32893531
#> 2     nec 0.09741274  2.27047288  2.17164742  2.32466685
#> 3     nec 0.16247747  2.26560937  2.16948326  2.31964916
#> 4     nec 0.22754220  2.26065387  2.16685457  2.31603919
#> 5     nec 0.29260692  2.25735357  2.16366209  2.31182104
#> 6     nec 0.35767165  2.25301312  2.15978555  2.30675523
#> 7     nec 0.42273638  2.24694088  2.15507938  2.30067007
#> 8     nec 0.48780111  2.23965822  2.14936746  2.29335922
#> 9     nec 0.55286583  2.23092491  2.14243701  2.28457489
#> 10    nec 0.61793056  2.22045395  2.13403125  2.27401984
#> 11    nec 0.68299529  2.20361141  2.12384082  2.26133799
#> 12    nec 0.74806002  2.18244467  2.11149368  2.24884826
#> 13    nec 0.81312474  2.15758293  2.09443551  2.23385978
#> 14    nec 0.87818947  2.13172393  2.06610066  2.21555756
#> 15    nec 0.94325420  2.09702192  2.03229578  2.19323230
#> 16    nec 1.00831893  2.05651464  1.99333176  2.16603434
#> 17    nec 1.07338365  2.01245067  1.94847657  2.13295155
#> 18    nec 1.13844838  1.96778793  1.89384086  2.09278651
#> 19    nec 1.20351311  1.91856444  1.83020910  2.04413468
#> 20    nec 1.26857784  1.86361958  1.75775770  1.98536609
#> 21    nec 1.33364256  1.79788589  1.67537165  1.91461447
#> 22    nec 1.39870729  1.71558798  1.58209411  1.83016589
#> 23    nec 1.46377202  1.61515410  1.47696405  1.73459220
#> 24    nec 1.52883675  1.49752210  1.35895739  1.62060916
#> 25    nec 1.59390147  1.36065150  1.22442876  1.48650811
#> 26    nec 1.65896620  1.20260977  1.06194741  1.32931281
#> 27    nec 1.72403093  1.02005182  0.88151214  1.14520000
#> 28    nec 1.78909566  0.81049621  0.68267293  0.93182846
#> 29    nec 1.85416038  0.59775380  0.46539930  0.69125886
#> 30    nec 1.91922511  0.34634467  0.23016636  0.42560383
#> 31    nec 1.98428984  0.03705918 -0.02661766  0.14801729
#> 32    nec 2.04935457 -0.25273347 -0.32887966 -0.09717714
#> 33    nec 2.11441929 -0.55698211 -0.69265275 -0.35678903
#> 34    nec 2.17948402 -0.89144140 -1.07667532 -0.62855844
#> 35    nec 2.24454875 -1.22206034 -1.47157361 -0.90423122
#> 36    nec 2.30961348 -1.55201150 -1.86532616 -1.17910054
#> 37    nec 2.37467820 -1.87537638 -2.24889209 -1.45646250
#> 38    nec 2.43974293 -2.19165437 -2.61888435 -1.73329712
#> 39    nec 2.50480766 -2.49171909 -2.96564342 -2.00662051
#> 40    nec 2.56987239 -2.77171347 -3.28286243 -2.27361063
#> 41    nec 2.63493712 -3.02895580 -3.56768526 -2.53171747
#> 42    nec 2.70000184 -3.26194198 -3.81916486 -2.77874954
#> 43    nec 2.76506657 -3.47024064 -4.03793854 -3.01293142
#> 44    nec 2.83013130 -3.66814834 -4.22582151 -3.23293123
#> 45    nec 2.89519603 -3.85569167 -4.38539814 -3.43785972
#> 46    nec 2.96026075 -4.02254066 -4.51966367 -3.62724566
#> 47    nec 3.02532548 -4.16987371 -4.63174177 -3.77970470
#> 48    nec 3.09039021 -4.29912780 -4.72468252 -3.90678154
#> 49    nec 3.15545494 -4.41188184 -4.80133232 -4.01045885
#> 50    nec 3.22051966 -4.49028696 -4.86426167 -4.10106849
#> 51    ecx 0.03234801  2.14617558  2.06963157  2.21348798
#> 52    ecx 0.09741274  2.14617558  2.06963157  2.21348798
#> 53    ecx 0.16247747  2.14617558  2.06963157  2.21348798
#> 54    ecx 0.22754220  2.14617558  2.06963157  2.21348798
#> 55    ecx 0.29260692  2.14617558  2.06963157  2.21348798
#> 56    ecx 0.35767165  2.14617558  2.06963157  2.21348798
#> 57    ecx 0.42273638  2.14617558  2.06963157  2.21348798
#> 58    ecx 0.48780111  2.14617558  2.06963157  2.21348798
#> 59    ecx 0.55286583  2.14617558  2.06963157  2.21348798
#> 60    ecx 0.61793056  2.14617558  2.06963157  2.21348798
#> 61    ecx 0.68299529  2.14617558  2.06963157  2.21348798
#> 62    ecx 0.74806002  2.14617558  2.06963157  2.21348798
#> 63    ecx 0.81312474  2.14617558  2.06963157  2.21348798
#> 64    ecx 0.87818947  2.14617558  2.06963157  2.21348798
#> 65    ecx 0.94325420  2.14617558  2.06963157  2.21348798
#> 66    ecx 1.00831893  2.14617558  2.06963157  2.21348798
#> 67    ecx 1.07338365  2.14617558  2.06963157  2.21348798
#> 68    ecx 1.13844838  2.14617558  2.06963157  2.21348798
#> 69    ecx 1.20351311  2.14617558  2.06963157  2.21348798
#> 70    ecx 1.26857784  2.14617558  2.06963157  2.21348798
#> 71    ecx 1.33364256  2.14617558  2.06963157  2.21348798
#> 72    ecx 1.39870729  2.14617558  2.06963157  2.21348798
#> 73    ecx 1.46377202  2.11269650  1.97628302  2.20309874
#> 74    ecx 1.52883675  1.87668598  1.65926933  1.96701488
#> 75    ecx 1.59390147  1.52581789  1.35084323  1.61939482
#> 76    ecx 1.65896620  1.18716587  1.04768437  1.28236839
#> 77    ecx 1.72403093  0.86030460  0.74238180  0.95548912
#> 78    ecx 1.78909566  0.54482356  0.44252756  0.63845145
#> 79    ecx 1.85416038  0.24032653  0.15369265  0.33095907
#> 80    ecx 1.91922511 -0.05356892 -0.12453561  0.03272456
#> 81    ecx 1.98428984 -0.32937570 -0.39672116 -0.25576515
#> 82    ecx 2.04935457 -0.59265997 -0.65921563 -0.53387958
#> 83    ecx 2.11441929 -0.84949872 -0.91197698 -0.80372256
#> 84    ecx 2.17948402 -1.09887078 -1.15537071 -1.05952051
#> 85    ecx 2.24454875 -1.35134721 -1.39483865 -1.29884451
#> 86    ecx 2.30961348 -1.59524565 -1.63467805 -1.52650563
#> 87    ecx 2.37467820 -1.82207851 -1.86847446 -1.74567727
#> 88    ecx 2.43974293 -2.04442163 -2.09509829 -1.95668305
#> 89    ecx 2.50480766 -2.26415260 -2.31479921 -2.15983404
#> 90    ecx 2.56987239 -2.47610045 -2.53083018 -2.35542926
#> 91    ecx 2.63493712 -2.68123319 -2.74044745 -2.54375616
#> 92    ecx 2.70000184 -2.87405510 -2.94384145 -2.72149975
#> 93    ecx 2.76506657 -3.05981600 -3.14119698 -2.89133767
#> 94    ecx 2.83013130 -3.23906990 -3.33269331 -3.05436722
#> 95    ecx 2.89519603 -3.41204473 -3.51850442 -3.21086137
#> 96    ecx 2.96026075 -3.57896044 -3.69879912 -3.36108212
#> 97    ecx 3.02532548 -3.74002929 -3.87374120 -3.50528097
#> 98    ecx 3.09039021 -3.89545609 -4.04348957 -3.64369934
#> 99    ecx 3.15545494 -4.04543847 -4.20922943 -3.77656899
#> 100   ecx 3.22051966 -4.19016715 -4.37055928 -3.90411234
#> 
#> $diff_list
#> $diff_list$`nec-ecx`
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,]  0.11942590  0.117734620  0.115662810  0.113125148  0.110017338
#>  [2,]  0.22280126  0.217784585  0.211911643  0.205037955  0.196995293
#>  [3,]  0.00949995  0.008468702  0.007183711  0.005582667  0.003588028
#>  [4,]  0.15529066  0.151734954  0.147513882  0.142503990  0.136559355
#>  [5,]  0.25608639  0.251470389  0.246042755  0.239662406  0.232164356
#>  [6,]  0.05459213  0.052762600  0.050542959  0.047850354  0.044584492
#>  [7,]  0.06057977  0.058268124  0.055468727  0.052079208  0.047975954
#>  [8,]  0.14954485  0.143541837  0.136614222  0.128621698  0.119403367
#>  [9,] -0.02287656 -0.024492703 -0.026465758 -0.028874288 -0.031814014
#> [10,]  0.09475353  0.093415141  0.091767259  0.089738514  0.087241168
#>               [,6]         [,7]         [,8]        [,9]       [,10]
#>  [1,]  0.106211956  0.101553441  0.095852046  0.08887658  0.08034572
#>  [2,]  0.187588002  0.176588840  0.163734313  0.14871948  0.13119222
#>  [3,]  0.001103332 -0.001991358 -0.005845073 -0.01064286 -0.01661427
#>  [4,]  0.129507663  0.121145691  0.111234112  0.09949157  0.08558797
#>  [5,]  0.223355904  0.213012313  0.200871927  0.18663073  0.16993635
#>  [6,]  0.040624043  0.035822337  0.030002229  0.02295000  0.01440815
#>  [7,]  0.043009822  0.037001042  0.029733175  0.02094600  0.01032717
#>  [8,]  0.108774957  0.096525723  0.082415066  0.06616884  0.04747537
#>  [9,] -0.035401508 -0.039778636 -0.045117913 -0.05162891 -0.05956590
#> [10,]  0.084167421  0.080384899  0.075731168  0.07000709  0.06296877
#>              [,11]        [,12]       [,13]         [,14]         [,15]
#>  [1,]  0.069917744  0.057178343  0.04162650  0.0226581707 -0.0004521413
#>  [2,]  0.110747027  0.086918379  0.05917383  0.0269070787 -0.0105687075
#>  [3,] -0.024043705 -0.033283025 -0.04476673 -0.0590301203 -0.0767307880
#>  [4,]  0.069137015  0.049687962  0.02671671 -0.0003835291 -0.0323117612
#>  [5,]  0.150381510  0.127497081  0.10074485  0.0695103045  0.0330957770
#>  [6,]  0.004066933 -0.008445520 -0.02357458 -0.0418520963 -0.0639109197
#>  [7,] -0.002497419 -0.017974738 -0.03663719 -0.0591165666 -0.0861591963
#>  [8,]  0.025981320  0.001287287 -0.02705652 -0.0595544960 -0.0967695173
#>  [9,] -0.069236964 -0.081014658 -0.09534853 -0.1127795675 -0.1339565614
#> [10,]  0.054317944  0.043690415  0.03064241  0.0146345637 -0.0049866512
#>              [,16]       [,17]       [,18]      [,19]      [,20]      [,21]
#>  [1,] -0.028571764 -0.06273162 -0.10414834 -0.1542450 -0.2146678 -0.2872945
#>  [2,] -0.054026754 -0.10433113 -0.16243889 -0.2293988 -0.3063451 -0.3944854
#>  [3,] -0.098673609 -0.12583926 -0.15941578 -0.2008321 -0.2517913 -0.3142997
#>  [4,] -0.069868267 -0.11396255 -0.16561905 -0.2259791 -0.2962973 -0.3779302
#>  [5,] -0.009285397 -0.05851301 -0.11556135 -0.1814964 -0.2574677 -0.3446918
#>  [6,] -0.090500807 -0.12250528 -0.16095866 -0.2070621 -0.2621965 -0.3279299
#>  [7,] -0.118641806 -0.15758728 -0.20417905 -0.2597721 -0.3258977 -0.4042590
#>  [8,] -0.139326028 -0.18791207 -0.24327963 -0.3062427 -0.3776722 -0.4584871
#>  [9,] -0.159654449 -0.19079412 -0.22846309 -0.2739357 -0.3286910 -0.3944250
#> [10,] -0.029009995 -0.05838289 -0.09423658 -0.1379117 -0.1909825 -0.2552756
#>            [,22]      [,23]      [,24]       [,25]        [,26]        [,27]
#>  [1,] -0.3742291 -0.4777762 -0.3325070 -0.13471362  0.028994146  0.156797201
#>  [2,] -0.4950795 -0.6094090 -0.5479218 -0.33830811 -0.158411589 -0.008607965
#>  [3,] -0.3906864 -0.4836042 -0.3772540 -0.16320061  0.014397432  0.152740902
#>  [4,] -0.4723139 -0.4195472 -0.2279084 -0.06188583  0.077526421  0.189605509
#>  [5,] -0.4444277 -0.5579390 -0.5078551 -0.30170627 -0.124812153  0.022608878
#>  [6,] -0.4060152 -0.4983735 -0.3628447 -0.13899280  0.052124090  0.209006906
#>  [7,] -0.4967108 -0.5419811 -0.3215881 -0.13147081  0.026897217  0.152387937
#>  [8,] -0.5496405 -0.5348693 -0.3289333 -0.14498609  0.016392784  0.154752914
#>  [9,] -0.4730532 -0.5666979 -0.4136983 -0.18645816  0.004257922  0.156786681
#> [10,] -0.3328798 -0.3507840 -0.1044506  0.10685058  0.280920163  0.415674997
#>           [,28]     [,29]     [,30]     [,31]     [,32]     [,33]       [,34]
#>  [1,] 0.2474013 0.3003200 0.3161642 0.2968944 0.2459773 0.1683919  0.07044647
#>  [2,] 0.1110255 0.2008025 0.2615155 0.2945078 0.3017192 0.2856942  0.24954425
#>  [3,] 0.2494523 0.3029583 0.3129263 0.2806947 0.2096123 0.1051810 -0.02508420
#>  [4,] 0.2740211 0.3309608 0.3612464 0.3664244 0.3488113 0.3114790  0.25816949
#>  [5,] 0.1406957 0.2300393 0.2917656 0.3275943 0.3398618 0.3314966  0.30594362
#>  [6,] 0.3304818 0.4159076 0.4654012 0.4800530 0.4620993 0.4150067  0.34343605
#>  [7,] 0.2444168 0.3031532 0.3297172 0.3263291 0.2963740 0.2443484  0.17567387
#>  [8,] 0.2698275 0.3615909 0.4303163 0.4766278 0.5015420 0.5064924  0.49333209
#>  [9,] 0.2697954 0.3425272 0.3750715 0.3686299 0.3257299 0.2503363  0.14781128
#> [10,] 0.5093867 0.5609776 0.5703544 0.5387383 0.4689327 0.3654584  0.23449228
#>             [,35]       [,36]       [,37]       [,38]       [,39]       [,40]
#>  [1,] -0.04059727 -0.15705276 -0.27140852 -0.37690005 -0.46796120 -0.54049510
#>  [2,]  0.19686092  0.13158368  0.05783149 -0.02028506 -0.09885199 -0.17432088
#>  [3,] -0.17211784 -0.32604558 -0.47708309 -0.61641857 -0.73691154 -0.83349781
#>  [4,]  0.19313826  0.12093854  0.04616934 -0.02678276 -0.09397665 -0.15209806
#>  [5,]  0.26703673  0.21883067  0.16540740  0.11067839  0.05820433  0.01105084
#>  [6,]  0.25306159  0.15025235  0.04165022 -0.06629519 -0.16776388 -0.25791112
#>  [7,]  0.09638519  0.01272846 -0.06927322 -0.14422182 -0.20765676 -0.25624203
#>  [8,]  0.46430957  0.42201634  0.36930799  0.30920306  0.24476758  0.17899507
#>  [9,]  0.02469602 -0.11167979 -0.25371361 -0.39409276 -0.52631258 -0.64507094
#> [10,]  0.08357475 -0.07889939 -0.24433825 -0.40464725 -0.55283927 -0.68346359
#>             [,41]       [,42]         [,43]       [,44]       [,45]       [,46]
#>  [1,] -0.59194917 -0.62121889 -0.6284309753 -0.61466404 -0.58165815 -0.53154993
#>  [2,] -0.24364250 -0.30435485 -0.3546220865 -0.39322775 -0.41953013 -0.43339073
#>  [3,] -0.90327443 -0.94531403 -0.9603023102 -0.95010049 -0.91731551 -0.86493085
#>  [4,] -0.19857322 -0.23161023 -0.2501751060 -0.25391660 -0.24305800 -0.21827344
#>  [5,] -0.02830771 -0.05803193 -0.0769117945 -0.08432025 -0.08013815 -0.06466288
#>  [6,] -0.33307991 -0.39086883 -0.4300708504 -0.45051763 -0.45287002 -0.43839257
#>  [7,] -0.28781985 -0.30134675 -0.2967445228 -0.27470341 -0.23647186 -0.18365882
#>  [8,]  0.11469242  0.05438130  0.0002223591 -0.04603334 -0.08306485 -0.10997589
#>  [9,] -0.74649166 -0.82816907 -0.8890606029 -0.92927479 -0.94980715 -0.95226944
#> [10,] -0.79280563 -0.87886448 -0.9411570785 -0.98041797 -0.99826200 -0.99686233
#>             [,47]        [,48]       [,49]        [,50]
#>  [1,] -0.46665377 -0.389295971 -0.30169962 -0.205912545
#>  [2,] -0.43508715 -0.425220736 -0.40462714 -0.374295446
#>  [3,] -0.79602151 -0.713554791 -0.62026609 -0.518593505
#>  [4,] -0.18056369 -0.131142064 -0.07133685 -0.002512902
#>  [5,] -0.03851198 -0.002530538  0.04229165  0.094889973
#>  [6,] -0.40874061 -0.365777776 -0.31143154 -0.247587512
#>  [7,] -0.11806430 -0.041545381  0.04408216  0.137110126
#>  [8,] -0.12626761 -0.131795489 -0.12671612 -0.111429588
#>  [9,] -0.93864406 -0.911081514 -0.87174661 -0.822711035
#> [10,] -0.97867621 -0.946232068 -0.90197850 -0.848187461
#> 
#> 
#> $diff_data
#>    comparison diff.Estimate    diff.Q2.5  diff.Q97.5          x
#> 1     nec-ecx   0.107089716 -0.015591847  0.24859724 0.03234801
#> 2     nec-ecx   0.105574880 -0.017076387  0.24389108 0.09741274
#> 3     nec-ecx   0.103715034 -0.018894628  0.23836325 0.16247747
#> 4     nec-ecx   0.101431831 -0.021121473  0.23187190 0.22754220
#> 5     nec-ecx   0.098629253 -0.023848555  0.22425132 0.29260692
#> 6     nec-ecx   0.095189689 -0.027187919  0.21530813 0.35767165
#> 7     nec-ecx   0.088455311 -0.031276499  0.20481703 0.42273638
#> 8     nec-ecx   0.079073117 -0.036281524  0.19251596 0.48780111
#> 9     nec-ecx   0.068087961 -0.042407047  0.17810070 0.55286583
#> 10    nec-ecx   0.055222071 -0.049901784  0.16121892 0.61793056
#> 11    nec-ecx   0.040149632 -0.059068481  0.14146375 0.68299529
#> 12    nec-ecx   0.022488851 -0.070275040  0.11836687 0.74806002
#> 13    nec-ecx   0.001571067 -0.083967628  0.09139137 0.81312474
#> 14    nec-ecx  -0.021117813 -0.100803926  0.05992458 0.87818947
#> 15    nec-ecx  -0.048111340 -0.125589477  0.02554750 0.94325420
#> 16    nec-ecx  -0.080184537 -0.155080554 -0.01362483 1.00831893
#> 17    nec-ecx  -0.118233916 -0.190145659 -0.05841217 1.07338365
#> 18    nec-ecx  -0.161698775 -0.239945906 -0.09646673 1.13844838
#> 19    nec-ecx  -0.216520589 -0.298973618 -0.14158671 1.20351311
#> 20    nec-ecx  -0.279246913 -0.366651432 -0.19631167 1.26857784
#> 21    nec-ecx  -0.361311024 -0.446285782 -0.26247985 1.33364256
#> 22    nec-ecx  -0.458370792 -0.537731299 -0.34218344 1.39870729
#> 23    nec-ecx  -0.516621398 -0.599799017 -0.36625571 1.46377202
#> 24    nec-ecx  -0.347675828 -0.538906820 -0.13222863 1.52883675
#> 25    nec-ecx  -0.141989445 -0.330072695  0.06888489 1.59390147
#> 26    nec-ecx   0.021645001 -0.150851716  0.23515657 1.65896620
#> 27    nec-ecx   0.155769797 -0.001584175  0.36917468 1.72403093
#> 28    nec-ecx   0.259623855  0.117701286  0.46913306 1.78909566
#> 29    nec-ecx   0.317056993  0.207380787  0.52833683 1.85416038
#> 30    nec-ecx   0.345481822  0.268321761  0.54673990 1.91922511
#> 31    nec-ecx   0.347009380  0.283802619  0.52553409 1.98428984
#> 32    nec-ecx   0.332795850  0.217794395  0.49420489 2.04935457
#> 33    nec-ecx   0.298586602  0.119403464  0.48590809 2.11441929
#> 34    nec-ecx   0.242018262 -0.003589798  0.45960548 2.17948402
#> 35    nec-ecx   0.144761723 -0.142525710  0.41992318 2.24454875
#> 36    nec-ecx   0.066833500 -0.288022194  0.37629956 2.30961348
#> 37    nec-ecx  -0.013811500 -0.430806310  0.32343036 2.37467820
#> 38    nec-ecx  -0.105258508 -0.568770024  0.26453501 2.43974293
#> 39    nec-ecx  -0.187710320 -0.695495278  0.20279085 2.50480766
#> 40    nec-ecx  -0.257076574 -0.799740110  0.14120762 2.56987239
#> 41    nec-ecx  -0.310449880 -0.878418952  0.08251739 2.63493712
#> 42    nec-ecx  -0.347611840 -0.930362878  0.02908832 2.70000184
#> 43    nec-ecx  -0.392346468 -0.955994633 -0.01713283 2.76506657
#> 44    nec-ecx  -0.421872691 -0.973596540 -0.05464789 2.83013130
#> 45    nec-ecx  -0.436200078 -0.987359656 -0.08079666 2.89519603
#> 46    nec-ecx  -0.435891648 -0.986828931 -0.07485831 2.96026075
#> 47    nec-ecx  -0.421913880 -0.969668977 -0.05641125 3.02532548
#> 48    nec-ecx  -0.377536874 -0.938323193 -0.01130888 3.09039021
#> 49    nec-ecx  -0.306565579 -0.895176324  0.04367930 3.15545494
#> 50    nec-ecx  -0.226750028 -0.842455265  0.12761059 3.22051966
#> 
#> $prob_diff
#>    comparison          x prob
#> 1     nec-ecx 0.03234801  0.9
#> 2     nec-ecx 0.09741274  0.9
#> 3     nec-ecx 0.16247747  0.9
#> 4     nec-ecx 0.22754220  0.9
#> 5     nec-ecx 0.29260692  0.9
#> 6     nec-ecx 0.35767165  0.9
#> 7     nec-ecx 0.42273638  0.8
#> 8     nec-ecx 0.48780111  0.8
#> 9     nec-ecx 0.55286583  0.8
#> 10    nec-ecx 0.61793056  0.8
#> 11    nec-ecx 0.68299529  0.7
#> 12    nec-ecx 0.74806002  0.6
#> 13    nec-ecx 0.81312474  0.5
#> 14    nec-ecx 0.87818947  0.4
#> 15    nec-ecx 0.94325420  0.1
#> 16    nec-ecx 1.00831893  0.0
#> 17    nec-ecx 1.07338365  0.0
#> 18    nec-ecx 1.13844838  0.0
#> 19    nec-ecx 1.20351311  0.0
#> 20    nec-ecx 1.26857784  0.0
#> 21    nec-ecx 1.33364256  0.0
#> 22    nec-ecx 1.39870729  0.0
#> 23    nec-ecx 1.46377202  0.0
#> 24    nec-ecx 1.52883675  0.0
#> 25    nec-ecx 1.59390147  0.1
#> 26    nec-ecx 1.65896620  0.8
#> 27    nec-ecx 1.72403093  0.9
#> 28    nec-ecx 1.78909566  1.0
#> 29    nec-ecx 1.85416038  1.0
#> 30    nec-ecx 1.91922511  1.0
#> 31    nec-ecx 1.98428984  1.0
#> 32    nec-ecx 2.04935457  1.0
#> 33    nec-ecx 2.11441929  1.0
#> 34    nec-ecx 2.17948402  0.9
#> 35    nec-ecx 2.24454875  0.8
#> 36    nec-ecx 2.30961348  0.6
#> 37    nec-ecx 2.37467820  0.5
#> 38    nec-ecx 2.43974293  0.2
#> 39    nec-ecx 2.50480766  0.2
#> 40    nec-ecx 2.56987239  0.2
#> 41    nec-ecx 2.63493712  0.1
#> 42    nec-ecx 2.70000184  0.1
#> 43    nec-ecx 2.76506657  0.1
#> 44    nec-ecx 2.83013130  0.0
#> 45    nec-ecx 2.89519603  0.0
#> 46    nec-ecx 2.96026075  0.0
#> 47    nec-ecx 3.02532548  0.0
#> 48    nec-ecx 3.09039021  0.0
#> 49    nec-ecx 3.15545494  0.2
#> 50    nec-ecx 3.22051966  0.2
#> 
# }