Posterior dispersion

Calculates posterior dispersion metric

dispersion(model, summary = FALSE)

Arguments

model
An object of class brmsfit whose distribution family is either gaussian, poisson or binomial.

summary
Logical. Should summary stats be returned instead of full vector? Defaults to FALSE.

Value

If summary is FALSE, an n-long numeric vector containing the dispersion metric, where n is the number of post warm-up posterior draws from the brmsfit object. If TRUE, then a data.frame containing the summary stats (mean, median, 95% highest density intervals). of the dispersion metric.

Details

This function calculates a dispersion metric which takes the ratio between the observed relative to simulated Pearson residuals sums of squares.

Examples

if (FALSE) {
library(brms)
set.seed(10)

dt <- data.frame(predictor = rpois(60, lambda = 15))
dt$response <- dt$predictor + round(rnorm(nrow(dt), 0, 1))
# mimic overdispersion
dt$response_od <- abs(dt$predictor + round(rnorm(nrow(dt), 0, 10)))
mod_pois <- brm(response ~ predictor, dt, family = poisson())
mod_pois_od <- brm(response_od~predictor, dt, family = poisson())

dispersion(mod_pois)
dispersion(mod_pois_od)

# compare with regular dispersion parameter
# from quasipoisson glm
dispersion(mod_pois, summary = TRUE)
summary(glm(response ~ predictor, data = dt, family = quasipoisson))
dispersion(mod_pois_od, summary = TRUE)
summary(glm(response_od ~ predictor, data = dt, family = quasipoisson))
}