ecx_manec

Extracts the predicted ECx value as desired from a "bayesmanecfit" object.

ecx_manec(
  object,
  ecx_val = 10,
  precision = 1000,
  posterior = FALSE,
  type = "absolute",
  hormesis_def = "control",
  xform = NA,
  x_range = NA,
  prob_vals = c(0.5, 0.025, 0.975)
)

Arguments

object
An object of class "bayesnecfit" or "bayesmanecfit" returned by bnec.

ecx_val
The desired percentage effect value. This must be a value between 1 and 99 (for type = "relative" and "absolute"), defaults to 10.

precision
The number of unique x values over which to find ECx - large values will make the ECx estimate more precise.

posterior
A logical value indicating if the full posterior sample of calculated ECx values should be returned instead of just the median and 95 credible intervals

type
A character vector, taking values of "relative", "absolute" (the default) or "direct". See Details.

hormesis_def
A character vector, taking values of "max" or "control". See Details.

xform
A function to apply to the returned estimated concentration values

x_range
A range of x values over which to consider extracting ECx

prob_vals
A vector indicating the probability values over which to return the estimated ECx value. Defaults to 0.5 (median) and 0.025 and 0.975 (95 percent credible intervals).

Value

A vector containing the estimated ECx value, including upper and lower 95% credible interval bounds.

Details

type "relative" is calculated as the percentage decrease from the maximum predicted value of the response (top) to the minimum predicted value of the response. Type "absolute" (the default) is calculated as the percentage decrease from the maximum value of the response (top) to 0 (or bot for a 4 parameter model fit). Type "direct" provides a direct estimate of the x value for a given y. Note that for the current version, ECx for an "nechorme" (NEC Hormesis) model is estimated at a percent decline from the control. For hormesis_def, if "max", then ECx values are calculated as a decline from the maximum estimates (i.e. the peak at nec); if "control", then ECx values are calculated relative to the control, which is assumed to be the lowest observed concentration.

See also