fit_bayesnec

Fits a concentration(dose)-response model using brms

fit_bayesnec(
  data,
  x_var,
  y_var,
  trials_var = NA,
  x_type = NA,
  y_type = NA,
  x_range = NA,
  precision = 1000,
  over_disp = FALSE,
  model = NA,
  sig_val = 0.01,
  ...
)

Arguments

data
A data.frame containing the data to use for the model.

x_var
A character indicating the column heading indicating the concentration (x) variable.

y_var
A character indicating the column heading indicating the response (y) variable.

trials_var
The column heading indicating the column for the number of "trials" for binomial response data. If not supplied, the model may run but will not be the model you intended!

x_type
The statistical distribution of the x (concentration) data. This will be guessed based on the characteristic of the input data if not supplied and will inform the priors used for estimating nec in NEC step models.

y_type
The statistical distribution (family) to use for the y (response) data.

x_range
A range of x values over which to consider extracting ECx.

precision
The length of the x vector used for posterior predictions, and over which to extract ecx values. Large values will be slower but more precise.

over_disp
If an over-dispersed model should be used. Only changes the model fit for "poisson" and "binomial" y_type data. For "poisson", a negative binomial model will be fit. For binomial a beta model will be fit.

model
A character vector indicating the model(s) to fit. See Details for more information.

sig_val
Probability value to use as the lower quantile to test significance of the predictor posterior values against the lowest observed concentration (assumed to be the control), to estimate NEC as an interpolated NOEC value from smooth ECX curves.

...
Further arguments to brm via fit_bayesnec.

Value

The fitted brms model, including an estimate of the NEC value and predicted posterior values. A posterior sample of the NEC is also available under nec_posterior

See also