Herbicide phytotoxicity data

Herbicide phytotoxicity dataset from Jones & Kerswell (2003).

Format

An object of class data.frame with 580 rows and 3 columns.

Details

The response data (Fv/Fm) Chlorophyll fluorescence measurements of symbiotic dinoflagellates still in the host tissue of the coral (in hospite or in vivo) were measured using a DIVING-PAM chlorophyll fluorometer (Walz) on vertical planes of tissue 2 to 3 cm above the base of the corals, using either a 6 mm (Acropora formosa) or 2 mm (Seriatopora hystrix) fibre-optic probe. Parameters measured were the maximum potential quantum yield (Fv/Fm).

Additional information on each of the herbicides included is available from the original publication Jones & Kerswell (2003).

The columns are as follows:

herbicide

The herbicide (chr).

concentration

The treatment concentration in µg / L (dbl).

fvfm

Maximum effective quantum yield (dbl).

References

Jones RJ, Kerswell AP (2003) Phytotoxicity of Photosystem II (PSII) herbicides to coral. Marine Ecology Progress Series, 261: 149-159. doi: 10.3354/meps261149.

Examples

head(herbicide)
#>     herbicide concentration fvfm
#> 1     irgarol           0.1 0.63
#> 2      diuron           0.1 0.68
#> 3     ametryn           0.1 0.68
#> 4 tebuthiuron           0.1 0.73
#> 5  hexazinone           0.1 0.64
#> 6    simazine           0.1 0.69