nec.bayesmanecfit

Extracts the predicted nec value as desired from an object of class bayesmanecfit.

# S3 method for bayesmanecfit
nec(object, posterior = FALSE, xform = NA, prob_vals = c(0.5, 0.025, 0.975))

Arguments

object
An object of class bayesmanecfit returned by bnec.

posterior
A logical value indicating if the full posterior sample of calculated nec values should be returned instead of just the median and 95 credible intervals.

xform
A function to apply to the returned estimated concentration values.

prob_vals
A vector indicating the probability values over which to return the estimated nec value. Defaults to 0.5 (median) and 0.025 and 0.975 (95 percent credible intervals).

Value

A vector containing the estimated nec value, including upper and lower 95% credible interval bounds (or other interval as specified by prob_vals).

See also

Examples

if (FALSE) {
library(brms)
library(bayesnec)
options(mc.cores = parallel::detectCores())
data(nec_data)

exmp <- bnec(data = nec_data, x_var = "x", y_var = "y",
             model = c("nec4param", "ecx4param"),
             family = Beta(link = "identity"), priors = my_priors,
             iter = 1e4, control = list(adapt_delta = 0.99))
exmp_2 <- pull_out(exmp, "ecx4param")
nec(exmp)
nec(exmp_2)
}