pull_out

Subsets model(s) from an existing object of class bayesmanecfit

pull_out(manec, model, loo_controls, ...)

Arguments

manec
An object of class bayesmanecfit as returned by bnec.

model
A character string indicating which model or suite of models to pull out.

loo_controls
A named list of two elements ("fitting" and/or "weights"), each being a named list containing the desired arguments to be passed on to loo (via "fitting") or to loo_model_weights (via "weights"). If "fitting" is provided with argument pointwise = TRUE (due to memory issues) and family = "beta_binomial2", the bnec will fail because that is a custom family. If "weights" is not provided by the user, bnec will set the default method argument in loo_model_weights to "pseudobma". See ?loo_model_weights for further info.

...
Additional arguments to expand_nec or expand_manec.

Value

If model is a string representing a single model, an object of class bayesnecfit; If model is instead a string depicting a suite of models, and object of class bayesmanecfit.

See also

Examples

library(bayesnec)
data(manec_example)
ecx4param <- pull_out(manec_example, model = "ecx4param")
#> Pulling out model(s): ecx4param
#> Warning: Found 1 observations with a pareto_k > 0.7 in model 'ecx4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 5 (5.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.
necmodels <- pull_out(manec_example, model = "nec")
#> Model(s) nec3param, nechorme, nechorme4, necsigm, neclin, neclinhorme, nechormepwr, nechorme4pwr, nechormepwr01 non-existent in current set of models: nec4param, ecx4param.
#> They may have been removed due to incompatibility with the modelled response distribution. If needed, add desired model(s) via function amend (see ?amend).
#> Pulling out model(s): nec4param
#> Warning: Found 2 observations with a pareto_k > 0.7 in model 'nec4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 4 (4.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.
#> Warning: Method 'posterior_samples' is deprecated. Please see ?as_draws for recommended alternatives.