pull_prior

Extracts the priors from an object of class bayesnecfit or bayesmanecfit.

pull_prior(object)

Arguments

object
An object of class bayesnecfit or bayesmanecfit returned by bnec.

Value

A list containing the priors

Examples

if (FALSE) {
library(brms)
library(bayesnec)
options(mc.cores = parallel::detectCores())
data(nec_data)

exmp_a <- bnec(data = nec_data, x_var = "x", y_var = "y",
               model = "nec3param",
               family = Beta(link = "identity"),
               iter = 1e4, control = list(adapt_delta = 0.99))
pull_prior(exmp_a)
}