Generates a summary for objects fitted by bnec

Generates a summary for objects fitted by bnec. object should be of class bayesnecfit or bayesmanecfit.

# S3 method for bayesnecfit
summary(object, ..., ecx = FALSE, ecx_vals = c(10, 50, 90))

# S3 method for bayesmanecfit
summary(object, ..., ecx = FALSE, ecx_vals = c(10, 50, 90))

Arguments

object

An object of class bayesnecfit or bayesmanecfit.

...

Unused.

ecx

Should summary ECx values be calculated? Defaults to FALSE.

ecx_vals

ECx targets (between 1 and 99). Only relevant if ecx = TRUE. If no value is specified by the user, returns calculations for EC10, EC50, and EC90.

Value

A summary of the fitted model. In the case of a bayesnecfit object, the summary contains most of the original contents of a brmsfit object with the addition of an R2. In the case of a bayesmanecfit object, summary displays the family distribution information, model weights and averaging method, the estimated model-averaged NEC, and R2 estimates for each individual model. Warning messages are also printed to screen in case model fits are not satisfactory with regards to their Rhats.

Examples

# \donttest{
library(bayesnec)
summary(manec_example)
#> Warning: Parts of the model have not converged (some Rhats are > 1.05). Be careful when analysing the results! We recommend running more iterations and/or setting stronger priors.
#> Object of class bayesmanecfit
#> 
#>  Family: gaussian  
#>   Links: mu = identity; sigma = identity  
#> 
#> Number of posterior draws per model:  100
#> 
#> Model weights (Method: pseudobma_bb_weights):
#>             waic   wi
#> nec4param 177.83 0.82
#> ecx4param 187.64 0.18
#> 
#> 
#> Summary of weighted NEC posterior estimates:
#> NB: Model set contains the ECX models: ecx4param; weighted NEC estimates include NSEC surrogates for NEC
#>     Estimate Q2.5 Q97.5
#> NEC     1.46 0.73  1.53
#> 
#> 
#> Bayesian R2 estimates:
#>           Estimate Est.Error Q2.5 Q97.5
#> nec4param     0.86      0.01 0.84  0.88
#> ecx4param     0.85      0.01 0.82  0.87
#> 
#> 
#> Warning: The following model had Rhats > 1.05 (no convergence):
#>   -  nec4param
#> Consider dropping them (see ?amend)
nec4param <- pull_out(manec_example, "nec4param")
#> Pulling out model(s): nec4param
summary(nec4param)
#> Warning: Parts of the model have not converged (some Rhats are > 1.05). Be careful when analysing the results! We recommend running more iterations and/or setting stronger priors.
#> Object of class bayesnecfit containing the following non-linear model: nec4param
#> 
#>  Family: gaussian 
#>   Links: mu = identity; sigma = identity 
#> Formula: y ~ bot + (top - bot) * exp(-exp(beta) * (x - nec) * step(x - nec)) 
#>          bot ~ 1
#>          top ~ 1
#>          beta ~ 1
#>          nec ~ 1
#>    Data: data (Number of observations: 100) 
#>   Draws: 2 chains, each with iter = 200; warmup = 150; thin = 1;
#>          total post-warmup draws = 100
#> 
#> Population-Level Effects: 
#>      Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> bot     -8.64      1.64   -13.13    -5.76 1.02       32       72
#> top      2.16      0.06     2.05     2.26 1.03       58       55
#> beta    -0.69      0.23    -1.14    -0.29 1.00       32       72
#> nec      1.47      0.03     1.41     1.53 1.01       65       80
#> 
#> Family Specific Parameters: 
#>       Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> sigma     0.52      0.03     0.47     0.59 1.06      101       72
#> 
#> Draws were sampled using sampling(NUTS). For each parameter, Bulk_ESS
#> and Tail_ESS are effective sample size measures, and Rhat is the potential
#> scale reduction factor on split chains (at convergence, Rhat = 1).
#> 
#> 
#> Bayesian R2 estimates:
#>    Estimate Est.Error Q2.5 Q97.5
#> R2     0.86      0.01 0.84  0.88
#> 
#> 
# }