Class bayesnecfit of models fitted with the brms package

Models fitted with the bayesnec package are represented as a bayesnecfit object, which contain the original brmsfit fitted object, list of initialisation values used, the validated bayesnecformula, name of non-linear model that was fitted, posterior predictions, posterior parameter estimates and a series of other statistics.

Details

See methods(class = "bayesnecfit") for an overview of available methods.

Slots

fit

The fitted Bayesian model of class brmsfit.

model

A character string indicating the name of the fitted model.

init

A list containing the initialisation values to fit the model.

bayesnecformula

An object of class bayesnecformula and formula.

pred_vals

A list containing a data.frame of summary posterior predicted values and a vector containing based on the supplied resolution and x_range.

top

The estimate for parameter "top" in the fitted model.

beta

The estimate for parameter "beta" in the fitted model.

ne

The estimated NEC.

f

The estimate for parameter "f" in the fitted model, NA if absent for the fitted model type.

bot

The estimate for parameter "bot" in the fitted model, NA if absent for the fitted model type.

d

The estimate for parameter "d" in the fitted model, NA if absent for the fitted model type.

slope

The estimate for parameter "slope" in the fitted model, NA if absent for the fitted model type.

ec50

The estimate for parameter "ec50" in the fitted model, NA if absent for the fitted model type.

dispersion

An estimate of dispersion.

predicted_y

The predicted values for the observed data.

residuals

Residual values of the observed data from the fitted model.

ne_posterior

A full posterior estimate of the NEC.

ne_type

A character vector indicating the type of no-effect toxicity estimate. Where the fitted model is an NEC model (threshold model, containing a step function) the no-effect estimate is a true no-effect-concentration (NEC, see Fox 2010). Where the fitted model is a smooth ECx model with no step function, the no-effect estimate is a no-significant-effect-concentration (NSEC, see Fisher and Fox 2023).

References

Fisher R, Fox DR (2023). Introducing the no significant effect concentration (NSEC).Environmental Toxicology and Chemistry, 42(9), 2019–2028. doi: 10.1002/etc.5610.

Fisher R, Fox DR, Negri AP, van Dam J, Flores F, Koppel D (2023). Methods for estimating no-effect toxicity concentrations in ecotoxicology. Integrated Environmental Assessment and Management. doi:10.1002/ieam.4809.

Fox DR (2010). A Bayesian Approach for Determining the No Effect Concentration and Hazardous Concentration in Ecotoxicology. Ecotoxicology and Environmental Safety, 73(2), 123–131. doi: 10.1016/j.ecoenv.2009.09.012.

See also