bnec

Fits a variety of NEC models using Bayesian analysis and provides a model averaged predictions based on WAIC model weights

bnec(
  data,
  x_var,
  y_var,
  model,
  trials_var = NA,
  family = NULL,
  priors,
  x_range = NA,
  precision = 1000,
  sig_val = 0.01,
  iter = 20000,
  warmup = floor(iter/5) * 4,
  inits,
  ...
)

Arguments

data
A data.frame containing the data to use for the model.

x_var
A character indicating the column heading containing the concentration (x) variable.

y_var
A character indicating the column heading containing the response (y) variable.

model
A character vector indicating the model(s) to fit. See Details for more information.

trials_var
The column heading indicating the column for the number of "trials" for binomial response data. If not supplied, the model may run but will not be the model you intended!

family
Either a character string, a function, or an object of class family defining the statistical distribution (family) to use for the y (response) data. See details.

priors
An object of class brmsprior which specifies user-desired prior distributions of model parameters. If missing, bnec will figure out a baseline prior for each parameter.

x_range
A range of x values over which to consider extracting ECx.

precision
The length of the x vector used for posterior predictions, and over which to extract ECx values. Large values will be slower but more precise.

sig_val
Probability value to use as the lower quantile to test significance of the predictor posterior values against the lowest observed concentration (assumed to be the control), to estimate NEC as an interpolated NOEC value from smooth ECx curves.

iter
The number of iterations to be passed to brm. Defaults to 2e3 to be consistent with brms defaults.

warmup
A positive integer specifying number of warmup (a.k.a. burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup samples should not be used for inference. The number of warmup should not be larger than "iter" and the default is "floor(iter / 5) * 4".

inits
Optional. Initialisation values. Must be a list of "n" names lists, where "n" corresponds to the number of chains, and names correspond to the parameter names of a given model.

...
Further arguments to brm via fit_bayesnec.

Value

If argument model is a single string, then an object of class bayesnecfit; if many strings or a set, an object of class bayesmanecfit.

Details

As some concentration-response data will use zero concentration which can cause numerical estimation issues, a small offset is added (1 / 10th of the next lowest value) to zero values of concentration where x_var are distributed on a continuous scale from 0 to infinity, or are bounded to 0, or 1.

The argument family indicates the family to use for the response variable in the brm call, and may currently be "Beta" / Beta / Beta(), "binomial" / binomial / binomial(), "Gamma" / Gamma / Gamma(), "gaussian" / gaussian / gaussian(), "negbinomial" / negbinomial / negbinomial(), or "poisson" / poisson / poisson(). Notice that families Beta and negbinomial are exported objects of package brms, so the user needs to load brms before calling these families. Other families can be added as required, please raise an issue on the GitHub development site if your required family is not currently available. If not supplied, the appropriate distribution will be guessed based on the characteristics of the input data through check_data. Guesses include all of the above families but negbinomial because this latter requires knowledge on whether the data is over-dispersed. As explained below in the Return section, the user can extract the dispersion parameter from a bnec call, and if they so wish, can refit the model using the negbinomial family.

The argument model may be one of "nec3param", "nec4param", "necsigm", "nechorme", "ecx4param", "ecxwb1", "ecxwb2", "ecxexp", "ecxlin", or "excsigm", in which case a single model of the specified type it fit, and bnec returns a model object of class bayesnecfit.

If a vector of two or more of the available models is supplied, bnec returns a model object of class bayesmanecfit containing model averaged predictions for the supplied models, providing they were successfully fitted.

Model averaging is achieved through a weighted sample of each fitted models posterior predictions, with weights derived using the loo_model_weights from brms. Individual model fits can be extracted from the mod_fits element and can be examined individually.

model may also be one of "all", meaning all of the available models will be fit; "ecx" meaning only models excluding a specific NEC step parameter fill be fit; "nec" meaning only models with a specific NEC step parameter will be fit; or "bot_free" meaning only models without a "bot" parameter (without a bottom plateau) will be fit.

Models are fitted using model formula passed to brms.

All models provide an estimate for NEC. For model types with "nec" as a prefix, NEC is directly estimated as parameter "nec" in the model. Models with "ecx" as a prefix are continuous curve models, typically used for extracting ECx values from concentration response data. In this instance the NEC value is defined as the concentration at which there is a user supplied (see sig_val) percentage certainty (based on the Bayesian posterior estimate) that the response falls below the estimated value of the upper asymptote (top) of the response (i.e. the response value is significantly lower than that expected in the case of no exposure). The default value for sig_val is 0.01, which corresponds to an alpha value of 0.01 for a one-sided test of significance.