bnec

Fits a variety of NEC models using Bayesian analysis and provides a model averaged predictions based on WAIC model weights

bnec(
  x,
  y = NULL,
  data,
  x_var,
  y_var,
  model,
  trials_var = NA,
  family = NULL,
  priors,
  x_range = NA,
  precision = 1000,
  sig_val = 0.01,
  iter = 10000,
  warmup = floor(iter/10) * 9,
  inits,
  sample_prior = "yes",
  loo_controls,
  random = NA,
  random_vars = NA,
  weights = NA,
  ...
)

Arguments

x
A numeric vector to use for the x-variable (typically concentration) in modelling, or a data.frame containing both the x and y data.

y
A numeric vector to use for the y-variable (typically the response) in modelling .

data
A data.frame containing the data to use for the model.

x_var
A character indicating the column heading containing the concentration (x) variable.

y_var
A character indicating the column heading containing the response (y) variable.

model
A character vector indicating the model(s) to fit. See Details for more information.

trials_var
A character indicating the column heading for the number of "trials" for binomial or beta_binomial2 response data, as it appears in "data" (if data is supplied). Alternatively a numeric vector representing the trials associated with each observation, if data are supplied as an x argument. or a numeric vector indicating the trials If not supplied, the model may run but will not be the model you intended!

family
Either a character string, a function, or an object of class family defining the statistical distribution (family) to use for the y (response) data. See details.

priors
An object of class brmsprior which specifies user-desired prior distributions of model parameters. If missing, bnec will figure out a baseline prior for each parameter. It can also be specified as a named list where each name needs to correspond to the same string as "model". See details.

x_range
A range of x values over which to consider extracting ECx.

precision
The length of the x vector used for posterior predictions, and over which to extract ECx values. Large values will be slower but more precise.

sig_val
Probability value to use as the lower quantile to test significance of the predicted posterior values against the lowest observed concentration (assumed to be the control), to estimate NEC as an interpolated NOEC value from smooth ECx curves.

iter
The number of iterations to be passed to brm. Defaults to 2e3 to be consistent with brms defaults.

warmup
A positive integer specifying number of warmup (a.k.a. burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup samples should not be used for inference. The number of warmup should not be larger than "iter" and the default is "floor(iter / 5) * 4".

inits
Optional. Initialisation values. Must be a list of "n" names lists, where "n" corresponds to the number of chains, and names correspond to the parameter names of a given model.

sample_prior
Indicate if samples from priors should be drawn additionally to the posterior samples. Options are "no", "yes" (the default), and "only". Among others, these samples can be used to calculate Bayes factors for point hypotheses via hypothesis.

loo_controls
A named list of two elements ("fitting" and/or "weights"), each being a named list containing the desired arguments to be passed on to loo (via "fitting") or to loo_model_weights (via "weights"). If "fitting" is provided with argument pointwise = TRUE (due to memory issues) and family = "beta_binomial2", the bnec will fail because that is a custom family. If "weights" is not provided by the user, bnec will set the default method argument in loo_model_weights to "pseudobma". See ?loo_model_weights for further info.

random
A named list containing the random model formula to apply to model parameters.

random_vars
A character vector containing the names of the columns containing the variables used in the random model formula.

weights
A character vector containing the name of the column which should be used as weights on the model likelihood. It must be a non-negative numeric column. See details.

...
Further arguments to brm via fit_bayesnec.

Value

If argument model is a single string, then an object of class bayesnecfit; if many strings or a set, an object of class bayesmanecfit.

Details

As some concentration-response data will use zero concentration which can cause numerical estimation issues, a small offset is added (1 / 10th of the next lowest value) to zero values of concentration where x_var are distributed on a continuous scale from 0 to infinity, or are bounded to 0, or 1.

The argument family indicates the family to use for the response variable in the brm call, and may currently be "Beta" / Beta / Beta(), "binomial" / binomial / binomial(), "beta_binomial2" / beta_binomial2, "Gamma" / Gamma / Gamma(), "gaussian" / gaussian / gaussian(), "negbinomial" / negbinomial / negbinomial(), or "poisson" / poisson / poisson(). Notice that families Beta and negbinomial are exported objects of package brms, so the user needs to load brms before calling these families.

Other families can be added as required, please raise an issue on the GitHub development site if your required family is not currently available.

If not supplied, the appropriate distribution will be guessed based on the characteristics of the input data through check_data. Guesses include all of the above families but "negbinomial" and "betabinomimal2" because these requires knowledge on whether the data is over-dispersed. As explained below in the Return section, the user can extract the dispersion parameter from a bnec call, and if they so wish, can refit the model using the "negbinomial" family.

The argument model may be a character string indicating the names of the desired model. see ?models for more details, and the list of models available. If a recongised model name is provided a single model of the specified type is fit, and bnec returns a model object of class bayesnecfit.

If a vector of two or more of the available models is supplied, bnec returns a model object of class bayesmanecfit containing model averaged predictions for the supplied models, providing they were successfully fitted.

Model averaging is achieved through a weighted sample of each fitted models posterior predictions, with weights derived using the loo_model_weights from brms. Individual model fits can be extracted from the mod_fits element and can be examined individually.

model may also be one of "all", meaning all of the available models will be fit; "ecx" meaning only models excluding a specific NEC step parameter will be fit; "nec" meaning only models with a specific NEC step parameter will be fit; or "bot_free" meaning only models without a "bot" parameter (without a bottom plateau) will be fit. Notice that if one of these group strings is provided together with a user-specified named list for the argument priors, the list names need to contain the actual model names, and not the group string , e.g. if model = "ecx" and priors = my_priors then names(my_priors) must #' contain models("ecx"). To check available models and associated parameters for each group, use the function models or to check the parameters of a specific model use the function show_params.

Models are fitted using model formula passed to brms.

All models provide an estimate for NEC. For model types with "nec" as a prefix, NEC is directly estimated as parameter "nec" in the model. Models with "ecx" as a prefix are continuous curve models, typically used for extracting ECx values from concentration response data. In this instance the NEC value is defined as the concentration at which there is a user supplied (see sig_val) percentage certainty (based on the Bayesian posterior estimate) that the response falls below the estimated value of the upper asymptote (top) of the response (i.e. the response value is significantly lower than that expected in the case of no exposure). The default value for sig_val is 0.01, which corresponds to an alpha value of 0.01 for a one-sided test of significance.

NAs are thrown away Stan's default behaviour is to fail when the input data contains NAs. For that reason brms excludes any NAs from input data prior to fitting, and does not allow them back in as is the case with e.g. stats::lm and na.action = exclude. So we advise that you exclude any NAs in your data prior to fitting because if you so wish that should facilitate merging predictions back onto your original dataset.

#' Weights Model weights are passed on to the model via an "aterm" formula element, e.g. y | weights(weight_vector), following brms formula syntax. Please note that brms does not implement design weights as in other standard base function. From their help page, brms "takes the weights literally, which means that an observation with weight 2 receives 2 times more weight than an observation with weight 1. It also means that using a weight of 2 is equivalent to adding the corresponding observation twice to the data frame."

Examples

# \donttest{
library(brms)
library(bayesnec)
data(nec_data)

# A single model
exmp_a <- bnec(data = nec_data, x_var = "x", y_var = "y",
               model = "nec4param", chains = 2)
#> Finding initial values which allow the response to be fitted using a nec4param model and a beta distribution.
#> Compiling Stan program...
#> Start sampling
#> Response variable modelled as a nec4param model using a beta distribution.
#> Warning: Found 1 observations with a pareto_k > 0.7 in model 'nec4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 2 (2.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.
#> Warning: Method 'posterior_samples' is deprecated. Please see ?as_draws for recommended alternatives.
# Two models model
exmp_b <- bnec(data = nec_data, x_var = "x", y_var = "y",
               model = c("nec4param", "ecx4param"), chains = 2)
#> Finding initial values which allow the response to be fitted using a nec4param model and a beta distribution.
#> Compiling Stan program...
#> Start sampling
#> Response variable modelled as a nec4param model using a beta distribution.
#> Finding initial values which allow the response to be fitted using a ecx4param model and a beta distribution.
#> Compiling Stan program...
#> Start sampling
#> Response variable modelled as a ecx4param model using a beta distribution.
#> Fitted models are:  nec4param ecx4param
#> Warning: Found 1 observations with a pareto_k > 0.7 in model 'nec4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 2 (2.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.
#> Warning: Method 'posterior_samples' is deprecated. Please see ?as_draws for recommended alternatives.
#> Warning: Found 1 observations with a pareto_k > 0.7 in model 'ecx4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 4 (4.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.
# }