compare_endpoints

Extracts posterior predicted values from a list of class bayesnecfit or bayesnecfit model fits and compares these via bootstrap re sampling.

compare_endpoints(
  x,
  comparison = "nec",
  ecx_val = 10,
  type = "absolute",
  hormesis_def = "control",
  sig_val = 0.01,
  precision,
  x_range = NA
)

Arguments

x
A named list list of objects of class bayesnecfit or bayesnecfit returned by bnec.

comparison
The posterior predictions to compare, takes values of "nec", "nsec", "ecx" or "fitted".

ecx_val
The desired percentage effect value. This must be a value between 1 and 99 (for type = "relative" and "absolute"), defaults to 10.

type
A character vector, taking values of "relative", "absolute" (the default) or "direct". See Details.

hormesis_def
A character vector, taking values of "max" or "control". See Details.

sig_val
Probability value to use as the lower quantile to test significance of the predicted posterior values. against the lowest observed concentration (assumed to be the control), to estimate NEC as an interpolated NOEC value from smooth ECx curves.

precision
The number of unique x values over which to find ECx - large values will make the ECx estimate more precise.

x_range
A range of x values over which to consider extracting ECx.

Value

A named list containing bootstrapped differences in posterior predictions of the bayesnecfit or bayesnecfit model fits contained in x. See Details.

See also

Examples

library(bayesnec)
manec_gauss_id_2 <- bayesnec:::manec_gauss_id_2
ecx4param <- pull_out(manec_gauss_id_2, model = "ecx4param")
#> Pulling out model(s): ecx4param
nec4param <- pull_out(manec_gauss_id_2, model = "nec4param")
#> Pulling out model(s): nec4param
compare_endpoints(list("nec" = ecx4param, "ecx" = nec4param), ecx_val = 50)
#> $posterior_list
#> $posterior_list$nec
#>  [1] 0.40254612 0.67062062 0.40573749 0.03553938 0.54296610 0.03234801
#>  [7] 0.52700928 0.51105247 0.55254018 0.65466380
#> 
#> $posterior_list$ecx
#>  [1] 1.375285 1.413673 1.508799 1.583274 1.516812 1.473901 1.444768 1.458241
#>  [9] 1.503794 1.499591
#> 
#> 
#> $posterior_data
#>    model      value
#> 1    ecx 1.58327384
#> 2    ecx 1.47390106
#> 3    ecx 1.49959058
#> 4    ecx 1.45824104
#> 5    ecx 1.51681216
#> 6    ecx 1.50379437
#> 7    ecx 1.41367264
#> 8    ecx 1.37528538
#> 9    ecx 1.50879934
#> 10   ecx 1.44476802
#> 11   nec 0.55254018
#> 12   nec 0.03553938
#> 13   nec 0.52700928
#> 14   nec 0.40573749
#> 15   nec 0.67062062
#> 16   nec 0.40254612
#> 17   nec 0.03234801
#> 18   nec 0.54296610
#> 19   nec 0.65466380
#> 20   nec 0.51105247
#> 
#> $diff_list
#> $diff_list$`nec-ecx`
#>  [1] -1.0307337 -1.4383617 -0.9725813 -1.0525036 -0.8461915 -1.1012482
#>  [7] -1.3813246 -0.8323193 -0.8541355 -0.9337156
#> 
#> 
#> $diff_data
#>    comparison       diff
#> 1     nec-ecx -1.0307337
#> 2     nec-ecx -1.4383617
#> 3     nec-ecx -0.9725813
#> 4     nec-ecx -1.0525036
#> 5     nec-ecx -0.8461915
#> 6     nec-ecx -1.1012482
#> 7     nec-ecx -1.3813246
#> 8     nec-ecx -0.8323193
#> 9     nec-ecx -0.8541355
#> 10    nec-ecx -0.9337156
#> 
#> $prob_diff
#>   comparison prob
#> 1    nec-ecx    0
#>