compare_fitted

Extracts posterior predicted values from a list of class bayesnecfit or bayesnecfit model fits and compares these across a vector of fitted values.

compare_fitted(x, precision = 50, x_range = NA)

Arguments

x
A named list list of objects of class bayesnecfit or bayesnecfit returned by bnec.

precision
The number of unique x values over which to find ECx - large values will make the ECx estimate more precise.

x_range
A range of x values over which to consider extracting ECx.

Value

A named list containing bootstrapped differences in posterior predictions of the bayesnecfit or bayesnecfit model fits contained in x. See Details.

See also

Examples

library(bayesnec)
manec_gauss_id_2 <- bayesnec:::manec_gauss_id_2
ecx4param <- pull_out(manec_gauss_id_2, model = "ecx4param")
#> Pulling out model(s): ecx4param
nec4param <- pull_out(manec_gauss_id_2, model = "nec4param")
#> Pulling out model(s): nec4param
compare_fitted(list("nec" = ecx4param, "ecx" = nec4param))
#> $posterior_list
#> $posterior_list$nec
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 2.235108 2.232543 2.229455 2.225739 2.221267 2.215888 2.209419 2.201642
#>  [2,] 2.294476 2.290917 2.286695 2.281690 2.275756 2.268724 2.260391 2.250523
#>  [3,] 2.244439 2.240739 2.236369 2.231209 2.225117 2.217926 2.209441 2.199433
#>  [4,] 2.211702 2.209132 2.206052 2.202358 2.197932 2.192628 2.186273 2.178663
#>  [5,] 2.248856 2.246579 2.243825 2.240496 2.236471 2.231606 2.225727 2.218626
#>  [6,] 2.211600 2.209468 2.206885 2.203757 2.199970 2.195386 2.189838 2.183126
#>  [7,] 2.270805 2.266480 2.261401 2.255438 2.248437 2.240222 2.230583 2.219280
#>  [8,] 2.273518 2.268721 2.263094 2.256493 2.248752 2.239678 2.229045 2.216589
#>  [9,] 2.273128 2.269048 2.264249 2.258606 2.251972 2.244175 2.235016 2.224261
#> [10,] 2.351677 2.343755 2.334727 2.324443 2.312728 2.299391 2.284211 2.266941
#>           [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>  [1,] 2.192296 2.181071 2.167596 2.151433 2.132060 2.108867 2.081131 2.048014
#>  [2,] 2.238841 2.225019 2.208675 2.189362 2.166563 2.139676 2.108004 2.070752
#>  [3,] 2.187634 2.173730 2.157355 2.138083 2.115421 2.088799 2.057559 2.020950
#>  [4,] 2.169553 2.158652 2.145614 2.130032 2.111422 2.089217 2.062752 2.031249
#>  [5,] 2.210051 2.199700 2.187213 2.172159 2.154022 2.132194 2.105951 2.074444
#>  [6,] 2.175008 2.165194 2.153335 2.139016 2.121738 2.100909 2.075827 2.045664
#>  [7,] 2.206030 2.190506 2.172329 2.151062 2.126199 2.097161 2.063286 2.023821
#>  [8,] 2.202007 2.184945 2.164994 2.141683 2.114474 2.082747 2.045799 2.002835
#>  [9,] 2.211638 2.196832 2.179477 2.159153 2.135374 2.107584 2.075152 2.037360
#> [10,] 2.247303 2.224984 2.199635 2.170864 2.138235 2.101266 2.059421 2.012114
#>          [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#>  [1,] 2.008540 1.961588 1.905879 1.839976 1.762286 1.671076 1.564510 1.440704
#>  [2,] 2.027008 1.975743 1.915802 1.845907 1.764661 1.670566 1.562052 1.437517
#>  [3,] 1.978116 1.928091 1.869791 1.802018 1.723460 1.632706 1.528269 1.408620
#>  [4,] 1.993810 1.949396 1.896824 1.834756 1.761700 1.676022 1.575962 1.459681
#>  [5,] 2.036676 1.991491 1.937558 1.873359 1.797192 1.707174 1.601277 1.477369
#>  [6,] 2.009448 1.966049 1.914160 1.852293 1.778768 1.691729 1.589166 1.468963
#>  [7,] 1.977915 1.924614 1.862857 1.791476 1.709203 1.614685 1.506502 1.383206
#>  [8,] 1.952959 1.895175 1.828383 1.751383 1.662885 1.561529 1.445915 1.314639
#>  [9,] 1.993404 1.942384 1.883313 1.815113 1.736633 1.646662 1.543965 1.427320
#> [10,] 1.958702 1.898486 1.830715 1.754585 1.669246 1.573812 1.467373 1.349013
#>          [,25]    [,26]     [,27]     [,28]     [,29]     [,30]        [,31]
#>  [1,] 1.297808 1.134123 0.9482390 0.7392104 0.5067370 0.2513434 -0.025474113
#>  [2,] 1.295399 1.134247 0.9528335 0.7502710 0.5261446 0.2806441  0.014680507
#>  [3,] 1.272242 1.117699 0.9437215 0.7493148 0.5338726 0.2973010  0.040132177
#>  [4,] 1.325320 1.171090 0.9953911 0.7969544 0.5750109 0.3294652  0.061059606
#>  [5,] 1.333301 1.167020 0.9767182 0.7610270 0.5192274 0.2514721 -0.041017898
#>  [6,] 1.328981 1.167165 0.9817033 0.7712115 0.5349501 0.2730475 -0.013300139
#>  [7,] 1.243367 1.085638 0.9088324 0.7120173 0.4946198 0.2565332 -0.001780801
#>  [8,] 1.166355 0.999842 0.8140890 0.6083935 0.3824634 0.1365190 -0.128618152
#>  [9,] 1.295572 1.147707 0.9829280 0.8007488 0.6010867 0.3843513  0.151517130
#> [10,] 1.217830 1.072966 0.9136404 0.7391865 0.5490952 0.3430598  0.121022313
#>             [,32]      [,33]      [,34]      [,35]     [,36]     [,37]
#>  [1,] -0.32116430 -0.6321289 -0.9538419 -1.2810820 -1.608256 -1.929776
#>  [2,] -0.27003134 -0.5709431 -0.8847118 -1.2073143 -1.534235 -1.860712
#>  [3,] -0.23638253 -0.5302199 -0.8385859 -1.1579823 -1.484341 -1.813218
#>  [4,] -0.22849887 -0.5364767 -0.8591418 -1.1918964 -1.529512 -1.866447
#>  [5,] -0.35581610 -0.6892530 -1.0365074 -1.3918383 -1.748941 -2.101383
#>  [6,] -0.32168884 -0.6484682 -0.9888335 -1.3370608 -1.686868 -2.031860
#>  [7,] -0.27920820 -0.5739317 -0.8834245 -1.2044993 -1.533416 -1.866041
#>  [8,] -0.41146202 -0.7098373 -1.0209040 -1.3412374 -1.666962 -1.993931
#>  [9,] -0.09583332 -0.3555066 -0.6247439 -0.9003182 -1.178678 -1.456126
#> [10,] -0.11678366 -0.3697955 -0.6370944 -0.9173920 -1.209035 -1.510027
#>           [,38]     [,39]     [,40]     [,41]     [,42]     [,43]     [,44]
#>  [1,] -2.240432 -2.535719 -2.812061 -3.066933 -3.298861 -3.507343 -3.692696
#>  [2,] -2.182006 -2.493677 -2.791815 -3.073219 -3.335494 -3.577078 -3.797201
#>  [3,] -2.140032 -2.460314 -2.769949 -3.065374 -3.343722 -3.602890 -3.841553
#>  [4,] -2.197202 -2.516678 -2.820468 -3.105078 -3.368024 -3.607844 -3.824014
#>  [5,] -2.443065 -2.768622 -3.073734 -3.355288 -3.611413 -3.841393 -4.045493
#>  [6,] -2.365987 -2.683953 -2.981520 -3.255669 -3.504621 -3.727744 -3.925379
#>  [7,] -2.198052 -2.525158 -2.843318 -3.148933 -3.438995 -3.711175 -3.963857
#>  [8,] -2.317934 -2.634912 -2.941155 -3.233465 -3.509268 -3.766672 -4.004468
#>  [9,] -1.729013 -1.993927 -2.247854 -2.488306 -2.713387 -2.921816 -3.112903
#> [10,] -1.818076 -2.130654 -2.445080 -2.758609 -3.068525 -3.372238 -3.667367
#>           [,45]     [,46]     [,47]     [,48]     [,49]     [,50]
#>  [1,] -3.855879 -3.998308 -4.121690 -4.227876 -4.318750 -4.396146
#>  [2,] -3.995796 -4.173374 -4.330896 -4.469641 -4.591085 -4.696806
#>  [3,] -4.059110 -4.255596 -4.431574 -4.588003 -4.726132 -4.847384
#>  [4,] -4.016809 -4.187137 -4.336361 -4.466141 -4.578292 -4.674675
#>  [5,] -4.224755 -4.380769 -4.515476 -4.630989 -4.729461 -4.812985
#>  [6,] -4.098620 -4.249096 -4.378763 -4.489734 -4.584149 -4.664079
#>  [7,] -4.196118 -4.407667 -4.598752 -4.770061 -4.922605 -5.057628
#>  [8,] -4.222091 -4.419543 -4.597297 -4.756195 -4.897344 -5.022026
#>  [9,] -3.286492 -3.442875 -3.582703 -3.706891 -3.816530 -3.912817
#> [10,] -3.951806 -4.223779 -4.481870 -4.725032 -4.952583 -5.164177
#> 
#> $posterior_list$ecx
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 2.098084 2.098084 2.098084 2.098084 2.098084 2.098084 2.098084 2.098084
#>  [2,] 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407
#>  [3,] 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824
#>  [4,] 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753
#>  [5,] 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898
#>  [6,] 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403
#>  [7,] 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292
#>  [8,] 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059
#>  [9,] 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100
#> [10,] 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015
#>           [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>  [1,] 2.098084 2.098084 2.098084 2.098084 2.098084 2.098084 2.098084 2.098084
#>  [2,] 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407
#>  [3,] 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824
#>  [4,] 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753
#>  [5,] 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898
#>  [6,] 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403
#>  [7,] 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292
#>  [8,] 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059
#>  [9,] 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100
#> [10,] 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015
#>          [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#>  [1,] 2.098084 2.098084 2.098084 2.098084 2.098084 2.003250 1.743527 1.489185
#>  [2,] 2.234407 2.234407 2.234407 2.234407 2.234407 2.234407 1.985938 1.670411
#>  [3,] 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824 2.122824 2.004049
#>  [4,] 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753 2.132753
#>  [5,] 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898 2.109898 2.038564
#>  [6,] 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403 2.156403 1.838424
#>  [7,] 2.162292 2.162292 2.162292 2.162292 2.162292 2.162292 2.069560 1.757569
#>  [8,] 2.143059 2.143059 2.143059 2.143059 2.143059 2.143059 2.114340 1.781134
#>  [9,] 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100 2.139100 2.001329
#> [10,] 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015 2.131015 1.972829
#>          [,25]     [,26]     [,27]     [,28]     [,29]       [,30]       [,31]
#>  [1,] 1.240113 0.9962002 0.7573408 0.5234300 0.2943651  0.07004591 -0.14962603
#>  [2,] 1.362791 1.0628801 0.7704850 0.4854175 0.2074937 -0.06346515 -0.32763376
#>  [3,] 1.627641 1.2650106 0.9156542 0.5790859 0.2548375 -0.05754185 -0.35848673
#>  [4,] 2.070495 1.6973891 1.3377631 0.9911299 0.6570199  0.33498080  0.02457645
#>  [5,] 1.660991 1.2972412 0.9468094 0.6092078 0.2839664 -0.02936707 -0.33122874
#>  [6,] 1.474818 1.1248099 0.7878916 0.4635735 0.1513845 -0.14912911 -0.43840395
#>  [7,] 1.453916 1.1583767 0.8707356 0.5907813 0.3183082  0.05311654 -0.20498836
#>  [8,] 1.456341 1.1397456 0.8311427 0.5303301 0.2371111 -0.04870605 -0.32730810
#>  [9,] 1.650013 1.3080786 0.9752761 0.6513615 0.3360974  0.02925270 -0.26939742
#> [10,] 1.628051 1.2929051 0.9671210 0.6504376 0.3426006  0.04336293 -0.24751562
#>            [,32]      [,33]      [,34]      [,35]     [,36]     [,37]     [,38]
#>  [1,] -0.3647470 -0.5754112 -0.7817110 -0.9837368 -1.181577 -1.375319 -1.565047
#>  [2,] -0.5851822 -0.8362765 -1.0810783 -1.3197453 -1.552431 -1.779286 -2.000456
#>  [3,] -0.6484156 -0.9277318 -1.1968238 -1.4560657 -1.705818 -1.946429 -2.178232
#>  [4,] -0.2746135 -0.5629942 -0.8409561 -1.1088757 -1.367116 -1.616026 -1.855944
#>  [5,] -0.6220386 -0.9022013 -1.1721065 -1.4321300 -1.682633 -1.923965 -2.156462
#>  [6,] -0.7168603 -0.9849029 -1.2429210 -1.4912896 -1.730370 -1.960508 -2.182040
#>  [7,] -0.4561959 -0.7006903 -0.9386510 -1.1702526 -1.395665 -1.615054 -1.828580
#>  [8,] -0.5988772 -0.8635910 -1.1216223 -1.3731401 -1.618309 -1.857288 -2.090235
#>  [9,] -0.5600718 -0.8429834 -1.1183396 -1.3863422 -1.647187 -1.901067 -2.148166
#> [10,] -0.5302686 -0.8051229 -1.0722993 -1.3320122 -1.584470 -1.829876 -2.068426
#>           [,39]     [,40]     [,41]     [,42]     [,43]     [,44]     [,45]
#>  [1,] -1.750844 -1.932792 -2.110971 -2.285458 -2.456330 -2.623662 -2.787528
#>  [2,] -2.216083 -2.426307 -2.631262 -2.831081 -3.025893 -3.215822 -3.400992
#>  [3,] -2.401550 -2.616693 -2.823961 -3.023643 -3.216015 -3.401345 -3.579891
#>  [4,] -2.087193 -2.310089 -2.524931 -2.732011 -2.931610 -3.123998 -3.309435
#>  [5,] -2.380446 -2.596229 -2.804112 -3.004384 -3.197324 -3.383200 -3.562270
#>  [6,] -2.395287 -2.600559 -2.798154 -2.988359 -3.171451 -3.347695 -3.517348
#>  [7,] -2.036400 -2.238666 -2.435527 -2.627127 -2.813607 -2.995104 -3.171750
#>  [8,] -2.317302 -2.538636 -2.754383 -2.964684 -3.169676 -3.369494 -3.564267
#>  [9,] -2.388666 -2.622743 -2.850569 -3.072311 -3.288130 -3.498187 -3.702633
#> [10,] -2.300312 -2.525721 -2.744833 -2.957824 -3.164866 -3.366123 -3.561759
#>           [,46]     [,47]     [,48]     [,49]     [,50]
#>  [1,] -2.947999 -3.105145 -3.259035 -3.409738 -3.557318
#>  [2,] -3.581522 -3.757527 -3.929122 -4.096416 -4.259518
#>  [3,] -3.751902 -3.917617 -4.077265 -4.231070 -4.379244
#>  [4,] -3.488172 -3.660452 -3.826508 -3.986564 -4.140838
#>  [5,] -3.734785 -3.900983 -4.061096 -4.215347 -4.363951
#>  [6,] -3.680657 -3.837858 -3.989179 -4.134842 -4.275057
#>  [7,] -3.343676 -3.511008 -3.673868 -3.832375 -3.986647
#>  [8,] -3.754124 -3.939187 -4.119580 -4.295418 -4.466818
#>  [9,] -3.901619 -4.095292 -4.283792 -4.467258 -4.645824
#> [10,] -3.751930 -3.936788 -4.116482 -4.291157 -4.460952
#> 
#> 
#> $posterior_data
#>     model          x      Estimate       Q2.5      Q97.5
#> 1     nec 0.03234801  2.259830e+00  2.2116230  2.3388068
#> 2     nec 0.09741274  2.256529e+00  2.2092079  2.3318664
#> 3     nec 0.16247747  2.252613e+00  2.2062390  2.3239203
#> 4     nec 0.22754220  2.247967e+00  2.2026730  2.3148232
#> 5     nec 0.29260692  2.242454e+00  2.1983904  2.3044097
#> 6     nec 0.35767165  2.235642e+00  2.1932482  2.2924908
#> 7     nec 0.42273638  2.227386e+00  2.1870753  2.2788515
#> 8     nec 0.48780111  2.217608e+00  2.1796673  2.2632470
#> 9     nec 0.55286583  2.204018e+00  2.1707803  2.2453990
#> 10    nec 0.61793056  2.187725e+00  2.1601236  2.2250110
#> 11    nec 0.68299529  2.169963e+00  2.1473513  2.2066406
#> 12    nec 0.74806002  2.151247e+00  2.1318432  2.1854916
#> 13    nec 0.81312474  2.129130e+00  2.1121085  2.1637417
#> 14    nec 0.87818947  2.101087e+00  2.0841085  2.1379924
#> 15    nec 0.94325420  2.069219e+00  2.0484450  2.1075425
#> 16    nec 1.00831893  2.034305e+00  2.0049225  2.0736131
#> 17    nec 1.07338365  1.993607e+00  1.9542512  2.0345006
#> 18    nec 1.13844838  1.945890e+00  1.8959204  1.9879478
#> 19    nec 1.20351311  1.890068e+00  1.8289079  1.9326629
#> 20    nec 1.26857784  1.824934e+00  1.7521031  1.8686193
#> 21    nec 1.33364256  1.749167e+00  1.6643158  1.7930462
#> 22    nec 1.39870729  1.658614e+00  1.5642927  1.7036990
#> 23    nec 1.46377202  1.553008e+00  1.4507428  1.5985517
#> 24    nec 1.52883675  1.432419e+00  1.3223731  1.4754776
#> 25    nec 1.59390147  1.295485e+00  1.1779369  1.3323292
#> 26    nec 1.65896620  1.134185e+00  1.0162949  1.1702071
#> 27    nec 1.72403093  9.505362e-01  0.8354063  0.9925869
#> 28    nec 1.78909566  7.497929e-01  0.6317089  0.7998951
#> 29    nec 1.85416038  5.300086e-01  0.4076986  0.5952196
#> 30    nec 1.91922511  2.768458e-01  0.1623545  0.3750607
#> 31    nec 1.98428984  6.449853e-03 -0.1089081  0.1446558
#> 32    nec 2.04935457 -2.746198e-01 -0.3989417 -0.1005472
#> 33    nec 2.11441929 -5.724374e-01 -0.7052058 -0.3587216
#> 34    nec 2.17948402 -8.840681e-01 -1.0329967 -0.6275227
#> 35    nec 2.24454875 -1.205907e+00 -1.3804531 -0.9041598
#> 36    nec 2.30961348 -1.533826e+00 -1.7349743 -1.1855084
#> 37    nec 2.37467820 -1.866244e+00 -2.0857405 -1.4682541
#> 38    nec 2.43974293 -2.197627e+00 -2.4257223 -1.7490525
#> 39    nec 2.50480766 -2.520918e+00 -2.7495717 -2.0246905
#> 40    nec 2.56987239 -2.816265e+00 -3.0529858 -2.2922303
#> 41    nec 2.63493712 -3.089148e+00 -3.3328736 -2.5491245
#> 42    nec 2.70000184 -3.355873e+00 -3.5884306 -2.7932931
#> 43    nec 2.76506657 -3.605367e+00 -3.8245803 -3.0231609
#> 44    nec 2.83013130 -3.832784e+00 -4.0362623 -3.2376576
#> 45    nec 2.89519603 -4.037960e+00 -4.2241553 -3.4146041
#> 46    nec 2.96026075 -4.236438e+00 -4.4168709 -3.5678476
#> 47    nec 3.02532548 -4.405168e+00 -4.5984249 -3.7039752
#> 48    nec 3.09039021 -4.538869e+00 -4.7669407 -3.8241123
#> 49    nec 3.15545494 -4.658609e+00 -4.9458379 -3.9295297
#> 50    nec 3.22051966 -4.754896e+00 -5.1402035 -4.0215661
#> 51    ecx 0.03234801  2.135926e+00  2.1007418  2.2181814
#> 52    ecx 0.09741274  2.135926e+00  2.1007418  2.2181814
#> 53    ecx 0.16247747  2.135926e+00  2.1007418  2.2181814
#> 54    ecx 0.22754220  2.135926e+00  2.1007418  2.2181814
#> 55    ecx 0.29260692  2.135926e+00  2.1007418  2.2181814
#> 56    ecx 0.35767165  2.135926e+00  2.1007418  2.2181814
#> 57    ecx 0.42273638  2.135926e+00  2.1007418  2.2181814
#> 58    ecx 0.48780111  2.135926e+00  2.1007418  2.2181814
#> 59    ecx 0.55286583  2.135926e+00  2.1007418  2.2181814
#> 60    ecx 0.61793056  2.135926e+00  2.1007418  2.2181814
#> 61    ecx 0.68299529  2.135926e+00  2.1007418  2.2181814
#> 62    ecx 0.74806002  2.135926e+00  2.1007418  2.2181814
#> 63    ecx 0.81312474  2.135926e+00  2.1007418  2.2181814
#> 64    ecx 0.87818947  2.135926e+00  2.1007418  2.2181814
#> 65    ecx 0.94325420  2.135926e+00  2.1007418  2.2181814
#> 66    ecx 1.00831893  2.135926e+00  2.1007418  2.2181814
#> 67    ecx 1.07338365  2.135926e+00  2.1007418  2.2181814
#> 68    ecx 1.13844838  2.135926e+00  2.1007418  2.2181814
#> 69    ecx 1.20351311  2.135926e+00  2.1007418  2.2181814
#> 70    ecx 1.26857784  2.135926e+00  2.1007418  2.2181814
#> 71    ecx 1.33364256  2.135926e+00  2.1007418  2.2181814
#> 72    ecx 1.39870729  2.135926e+00  2.0272458  2.2181814
#> 73    ecx 1.46377202  2.118582e+00  1.7980697  2.1525099
#> 74    ecx 1.52883675  1.905627e+00  1.5299611  2.1115602
#> 75    ecx 1.59390147  1.551229e+00  1.2677153  1.9783564
#> 76    ecx 1.65896620  1.211694e+00  1.0112031  1.6097942
#> 77    ecx 1.72403093  8.931949e-01  0.7602983  1.2562036
#> 78    ecx 1.78909566  5.849336e-01  0.4684884  0.9146820
#> 79    ecx 1.85416038  2.891658e-01  0.1640091  0.5862756
#> 80    ecx 1.91922511 -5.718427e-05 -0.1298547  0.2753705
#> 81    ecx 1.98428984 -2.983528e-01 -0.4204226 -0.0146191
#> 82    ecx 2.04935457 -5.726270e-01 -0.7014603 -0.2948935
#> 83    ecx 2.11441929 -8.396300e-01 -0.9720394 -0.5657880
#> 84    ecx 2.17948402 -1.099709e+00 -1.2325491 -0.7950411
#> 85    ecx 2.24454875 -1.352576e+00 -1.4833642 -1.0118931
#> 86    ecx 2.30961348 -1.601389e+00 -1.7248456 -1.2233234
#> 87    ecx 2.37467820 -1.843582e+00 -1.9573404 -1.4292593
#> 88    ecx 2.43974293 -2.079330e+00 -2.1811833 -1.6243417
#> 89    ecx 2.50480766 -2.308807e+00 -2.4001406 -1.8150940
#> 90    ecx 2.56987239 -2.532179e+00 -2.6213817 -2.0016137
#> 91    ecx 2.63493712 -2.749608e+00 -2.8445822 -2.1839958
#> 92    ecx 2.70000184 -2.961254e+00 -3.0613603 -2.3623334
#> 93    ecx 2.76506657 -3.167271e+00 -3.2719044 -2.5367174
#> 94    ecx 2.83013130 -3.356909e+00 -3.4763972 -2.7072366
#> 95    ecx 2.89519603 -3.539554e+00 -3.6750161 -2.8739779
#> 96    ecx 2.96026075 -3.707721e+00 -3.8684328 -3.0370261
#> 97    ecx 3.02532548 -3.869420e+00 -4.0601683 -3.1964640
#> 98    ecx 3.09039021 -4.025138e+00 -4.2468441 -3.3523726
#> 99    ecx 3.15545494 -4.175095e+00 -4.4285939 -3.5048312
#> 100   ecx 3.22051966 -4.319504e+00 -4.6055478 -3.6539171
#> 
#> $diff_list
#> $diff_list$`nec-ecx`
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,] -0.02270533 -0.02527500 -0.02835590 -0.03204917 -0.03647572 -0.04177999
#>  [2,]  0.13006835  0.12598844  0.12118965  0.11554662  0.10891265  0.10111620
#>  [3,]  0.14250323  0.13770624  0.13207843  0.12547737  0.11773680  0.10866280
#>  [4,]  0.19527390  0.18735195  0.17832435  0.16803938  0.15632529  0.14298786
#>  [5,]  0.04930846  0.04717592  0.04459313  0.04146545  0.03767855  0.03309440
#>  [6,]  0.13454100  0.13084113  0.12647124  0.12131095  0.11521867  0.10802793
#>  [7,]  0.17272133  0.16839665  0.16331773  0.15735422  0.15035368  0.14213806
#>  [8,]  0.15537620  0.15181688  0.14759560  0.14259020  0.13665636  0.12962375
#>  [9,]  0.10235565  0.09979029  0.09670240  0.09298617  0.08851463  0.08313553
#> [10,]  0.12603149  0.12375445  0.12100085  0.11767142  0.11364642  0.10878154
#>              [,7]        [,8]        [,9]        [,10]        [,11]       [,12]
#>  [1,] -0.04813434 -0.05574428 -0.06485448 -0.075755816 -0.088793417 -0.10437589
#>  [2,]  0.09195703  0.08120170  0.06857859  0.053772357  0.036417843  0.01609348
#>  [3,]  0.09802949  0.08557420  0.07099196  0.053929454  0.033978369  0.01066819
#>  [4,]  0.12780779  0.11053784  0.09089978  0.068581138  0.043231836  0.01446073
#>  [5,]  0.02754652  0.02083427  0.01271618  0.002902016 -0.008956392 -0.02327584
#>  [6,]  0.09954330  0.08953558  0.07773643  0.063832179  0.047457014  0.02818534
#>  [7,]  0.13249959  0.12119616  0.10794608  0.092422224  0.074245562  0.05297804
#>  [8,]  0.12129157  0.11142337  0.09974122  0.085918943  0.069574703  0.05026260
#>  [9,]  0.07666652  0.06888945  0.05954367  0.048318332  0.034843478  0.01867988
#> [10,]  0.10290299  0.09580170  0.08722645  0.076875838  0.064388922  0.04933440
#>               [,13]         [,14]       [,15]       [,16]       [,17]
#>  [1,] -0.1229856746 -0.1451905119 -0.17165590 -0.20315819 -0.24059780
#>  [2,] -0.0076856517 -0.0354749116 -0.06790710 -0.10569880 -0.14965557
#>  [3,] -0.0165414145 -0.0482684011 -0.08521611 -0.12818048 -0.17805599
#>  [4,] -0.0181678138 -0.0551372566 -0.09698185 -0.14428912 -0.19770156
#>  [5,] -0.0405539525 -0.0613829522 -0.08646487 -0.11662795 -0.15284357
#>  [6,]  0.0055235078 -0.0210990308 -0.05233890 -0.08894795 -0.13178184
#>  [7,]  0.0281149789 -0.0009229660 -0.03479794 -0.07426284 -0.12016849
#>  [8,]  0.0274636191  0.0005758628 -0.03109548 -0.06834815 -0.11209214
#>  [9,] -0.0006923397 -0.0238862422 -0.05162158 -0.08473879 -0.12421274
#> [10,]  0.0311982864  0.0093700454 -0.01687265 -0.04838025 -0.08614813
#>            [,18]      [,19]      [,20]      [,21]      [,22]      [,23]
#>  [1,] -0.2850117 -0.3375839 -0.3996518 -0.4727070 -0.5583855 -0.4099758
#>  [2,] -0.2006749 -0.2597462 -0.3279460 -0.4064265 -0.4963970 -0.5703743
#>  [3,] -0.2358397 -0.3026319 -0.3796326 -0.4681306 -0.5694860 -0.6851005
#>  [4,] -0.2579171 -0.3256882 -0.4018185 -0.4871575 -0.5825914 -0.6890301
#>  [5,] -0.1962429 -0.2481312 -0.3099989 -0.3835238 -0.4705627 -0.4803942
#>  [6,] -0.1818073 -0.2401068 -0.3078798 -0.3864378 -0.4771916 -0.5816291
#>  [7,] -0.1734693 -0.2352266 -0.3066077 -0.3888802 -0.3885652 -0.2370255
#>  [8,] -0.1633572 -0.2232978 -0.2931927 -0.3744386 -0.4685335 -0.5770483
#>  [9,] -0.1711651 -0.2268737 -0.2927766 -0.3704671 -0.4616771 -0.5682430
#> [10,] -0.1313328 -0.1852662 -0.2494649 -0.3256326 -0.4156499 -0.5215474
#>            [,24]        [,25]        [,26]        [,27]       [,28]      [,29]
#>  [1,] -0.2107298 -0.037471033  0.108210157  0.224906084  0.31153693  0.3675171
#>  [2,] -0.3538148 -0.160768595  0.007961216  0.151785265  0.27041875  0.3639756
#>  [3,] -0.6581900 -0.461696269 -0.293063141 -0.153031972 -0.04204406  0.0398628
#>  [4,] -0.4894118 -0.256988479 -0.051844045  0.125748769  0.27561299  0.3977107
#>  [5,] -0.2886059 -0.124934923  0.008788303  0.110967694  0.18043023  0.2166419
#>  [6,] -0.6299440 -0.388748703 -0.179542532 -0.003087878  0.14010707  0.2499062
#>  [7,] -0.1059795  0.003254672  0.089438256  0.151491572  0.18858731  0.2002546
#>  [8,] -0.5638116 -0.354613943 -0.173831684 -0.022442653  0.09890946  0.1900472
#>  [9,] -0.6920490 -0.772686448 -0.563266246 -0.389524185 -0.25191949 -0.1502829
#> [10,] -0.5266804 -0.294339409 -0.097990697  0.061064010  0.18194109  0.2643899
#>             [,30]       [,31]       [,32]        [,33]       [,34]        [,35]
#>  [1,]  0.39293032  0.38869336  0.35668337  0.299799801  0.22193647  0.127848881
#>  [2,]  0.43305734  0.47882523  0.50304390  0.508084369  0.49687849  0.472821902
#>  [3,]  0.09315606  0.11889747  0.11880656  0.095285602  0.05139526 -0.009225181
#>  [4,]  0.49218892  0.55942627  0.60007668  0.615107366  0.60582661  0.573897605
#>  [5,]  0.21993093  0.19168822  0.13450702  0.052222065 -0.05018251 -0.166808180
#>  [6,]  0.32666803  0.37136092  0.38565607  0.371981337  0.33352060  0.274147712
#>  [7,]  0.18648725  0.14784523  0.08553876  0.001479464 -0.10171353 -0.220762492
#>  [8,]  0.25139139  0.28407793  0.29004046  0.272040298  0.23362786  0.179027892
#>  [9,] -0.08363737 -0.05005057 -0.04655082 -0.069134697 -0.11288584 -0.172206310
#> [10,]  0.30901395  0.31746883  0.29259955  0.238478848  0.16031636  0.064227478
#>             [,36]       [,37]       [,38]       [,39]      [,40]       [,41]
#>  [1,]  0.02291905 -0.08716077 -0.19674649 -0.30059460 -0.3941616 -0.47381566
#>  [2,]  0.43963039  0.40116174  0.36122160  0.32337467  0.2907815  0.26607672
#>  [3,] -0.08249184 -0.16405534 -0.24950838 -0.33460009 -0.4154344 -0.48863218
#>  [4,]  0.52133495  0.45048148  0.36396454  0.26463310  0.1554787  0.03954514
#>  [5,] -0.29120336 -0.41680664 -0.53740720 -0.64755354 -0.7428545 -0.82014207
#>  [6,]  0.19829264  0.11074737  0.01642983 -0.07986808 -0.1737195 -0.26126183
#>  [7,] -0.35183837 -0.49072188 -0.63300517 -0.77431387 -0.9105258 -1.03796270
#>  [8,]  0.11295202  0.04035474 -0.03384034 -0.10501110 -0.1690722 -0.22265010
#>  [9,] -0.24114028 -0.31374951 -0.38448826 -0.44852521 -0.5019725 -0.54200180
#> [10,] -0.04312229 -0.15495440 -0.26483310 -0.36707268 -0.4570408 -0.53132654
#>             [,42]      [,43]      [,44]      [,45]      [,46]      [,47]
#>  [1,] -0.53694278 -0.5819514 -0.6081917 -0.6158172 -0.6056155 -0.5788343
#>  [2,]  0.25129704  0.2478604  0.2565904  0.2777748  0.3112482  0.3564843
#>  [3,] -0.55144384 -0.6018059 -0.6383443 -0.6603319 -0.6676135 -0.6605094
#>  [4,] -0.08016584 -0.2007875 -0.3196716 -0.4344573 -0.5431221 -0.6440123
#>  [5,] -0.87749349 -0.9141368 -0.9302751 -0.9268701 -0.9054199 -0.8677548
#>  [6,] -0.33933737 -0.4055662 -0.4583533 -0.4968394 -0.5208119 -0.5305908
#>  [7,] -1.15353759 -1.2548451 -1.3401947 -1.4085902 -1.4596681 -1.4936075
#>  [8,] -0.26318321 -0.2889474 -0.2990146 -0.2931627 -0.2717545 -0.2356047
#>  [9,] -0.56685003 -0.5757332 -0.5686983 -0.5464440 -0.5101361 -0.4612383
#> [10,] -0.58777063 -0.6253779 -0.6441481 -0.6448633 -0.6288670 -0.5978594
#>            [,48]      [,49]       [,50]
#>  [1,] -0.5370195 -0.4818758 -0.41515689
#>  [2,]  0.4126890  0.4788877  0.55400041
#>  [3,] -0.6397126 -0.6061868 -0.56107408
#>  [4,] -0.7358531 -0.8177409 -0.88912005
#>  [5,] -0.8158662 -0.7517741 -0.67743177
#>  [6,] -0.5269073 -0.5107848 -0.48343265
#>  [7,] -1.5110251 -1.5128677 -1.50031016
#>  [8,] -0.1858492 -0.1238273 -0.05098183
#>  [9,] -0.4013685 -0.3321859 -0.25530807
#> [10,] -0.5537239 -0.4983913 -0.43374072
#> 
#> 
#> $diff_data
#>    comparison diff.Estimate    diff.Q2.5   diff.Q97.5          x
#> 1     nec-ecx   0.132304679 -0.006502227  0.190199569 0.03234801
#> 2     nec-ecx   0.128414783 -0.008973543  0.183087008 0.09741274
#> 3     nec-ecx   0.123830441 -0.011942367  0.174947862 0.16247747
#> 4     nec-ecx   0.119491186 -0.015508380  0.165635221 0.22754220
#> 5     nec-ecx   0.114432545 -0.019791012  0.154981678 0.29260692
#> 6     nec-ecx   0.108345365 -0.024933251  0.142796655 0.35767165
#> 7     nec-ecx   0.098786395 -0.031106147  0.131443937 0.42273638
#> 8     nec-ecx   0.087554891 -0.038514106  0.118997285 0.48780111
#> 9     nec-ecx   0.074364193 -0.047401084  0.106099985 0.55286583
#> 10    nec-ecx   0.058880816 -0.058057804  0.090958986 0.61793056
#> 11    nec-ecx   0.039824839 -0.070830086  0.073194619 0.68299529
#> 12    nec-ecx   0.017386677 -0.086128379  0.052367069 0.74806002
#> 13    nec-ecx  -0.004188996 -0.104438537  0.030504542 0.81312474
#> 14    nec-ecx  -0.029680577 -0.126333811  0.007391354 0.87818947
#> 15    nec-ecx  -0.060122997 -0.154854238 -0.020072785 0.94325420
#> 16    nec-ecx  -0.097323373 -0.189912649 -0.052873027 1.00831893
#> 17    nec-ecx  -0.140718700 -0.230946148 -0.091985529 1.07338365
#> 18    nec-ecx  -0.189025063 -0.278915427 -0.138538321 1.13844838
#> 19    nec-ecx  -0.244119031 -0.334907343 -0.193823282 1.20351311
#> 20    nec-ecx  -0.308939378 -0.401331011 -0.259209999 1.26857784
#> 21    nec-ecx  -0.387659004 -0.483906158 -0.335720332 1.33364256
#> 22    nec-ecx  -0.473877189 -0.579642661 -0.394659235 1.39870729
#> 23    nec-ecx  -0.569308615 -0.688145931 -0.275939308 1.46377202
#> 24    nec-ecx  -0.508046112 -0.684430709 -0.129548284 1.52883675
#> 25    nec-ecx  -0.275663944 -0.702713658 -0.005908612 1.59390147
#> 26    nec-ecx  -0.074917371 -0.502470548  0.103986479 1.65896620
#> 27    nec-ecx   0.086015852 -0.336313437  0.208453900 1.72403093
#> 28    nec-ecx   0.181185662 -0.204697521  0.303454047 1.78909566
#> 29    nec-ecx   0.233274054 -0.107500135  0.390917129 1.85416038
#> 30    nec-ecx   0.280202672 -0.043858846  0.478884314 1.91922511
#> 31    nec-ecx   0.300773379 -0.012037259  0.541291032 1.98428984
#> 32    nec-ecx   0.291320001 -0.016830665  0.578244301 2.04935457
#> 33    nec-ecx   0.255259573 -0.053246511  0.591027192 2.11441929
#> 34    nec-ecx   0.191126413 -0.110372068  0.581313283 2.17948402
#> 35    nec-ecx   0.096038179 -0.209837351  0.551155571 2.24454875
#> 36    nec-ecx  -0.010101621 -0.338195492  0.502951428 2.30961348
#> 37    nec-ecx  -0.121057586 -0.474090954  0.439384537 2.37467820
#> 38    nec-ecx  -0.223127433 -0.611495628  0.363347379 2.43974293
#> 39    nec-ecx  -0.317597346 -0.745792796  0.310157818 2.50480766
#> 40    nec-ecx  -0.404798014 -0.872799730  0.260338373 2.56987239
#> 41    nec-ecx  -0.481223919 -0.988953061  0.215107115 2.63493712
#> 42    nec-ecx  -0.544193308 -1.091427665  0.176717890 2.70000184
#> 43    nec-ecx  -0.578842292 -1.178185769  0.146914648 2.76506657
#> 44    nec-ecx  -0.588445025 -1.247962798  0.131579245 2.83013130
#> 45    nec-ecx  -0.581130573 -1.300203166  0.149313862 2.89519603
#> 46    nec-ecx  -0.574368791 -1.334962253  0.180072625 2.96026075
#> 47    nec-ecx  -0.588346890 -1.352790636  0.223264254 3.02532548
#> 48    nec-ecx  -0.545371720 -1.354614363  0.278017875 3.09039021
#> 49    nec-ecx  -0.504588042 -1.356464144  0.343276793 3.15545494
#> 50    nec-ecx  -0.458586687 -1.362792385  0.417879404 3.22051966
#> 
#> $prob_diff
#>    comparison          x prob
#> 1     nec-ecx 0.03234801  0.9
#> 2     nec-ecx 0.09741274  0.9
#> 3     nec-ecx 0.16247747  0.9
#> 4     nec-ecx 0.22754220  0.9
#> 5     nec-ecx 0.29260692  0.9
#> 6     nec-ecx 0.35767165  0.9
#> 7     nec-ecx 0.42273638  0.9
#> 8     nec-ecx 0.48780111  0.9
#> 9     nec-ecx 0.55286583  0.9
#> 10    nec-ecx 0.61793056  0.9
#> 11    nec-ecx 0.68299529  0.8
#> 12    nec-ecx 0.74806002  0.8
#> 13    nec-ecx 0.81312474  0.4
#> 14    nec-ecx 0.87818947  0.2
#> 15    nec-ecx 0.94325420  0.0
#> 16    nec-ecx 1.00831893  0.0
#> 17    nec-ecx 1.07338365  0.0
#> 18    nec-ecx 1.13844838  0.0
#> 19    nec-ecx 1.20351311  0.0
#> 20    nec-ecx 1.26857784  0.0
#> 21    nec-ecx 1.33364256  0.0
#> 22    nec-ecx 1.39870729  0.0
#> 23    nec-ecx 1.46377202  0.0
#> 24    nec-ecx 1.52883675  0.0
#> 25    nec-ecx 1.59390147  0.1
#> 26    nec-ecx 1.65896620  0.4
#> 27    nec-ecx 1.72403093  0.6
#> 28    nec-ecx 1.78909566  0.8
#> 29    nec-ecx 1.85416038  0.9
#> 30    nec-ecx 1.91922511  0.9
#> 31    nec-ecx 1.98428984  0.9
#> 32    nec-ecx 2.04935457  0.9
#> 33    nec-ecx 2.11441929  0.9
#> 34    nec-ecx 2.17948402  0.7
#> 35    nec-ecx 2.24454875  0.6
#> 36    nec-ecx 2.30961348  0.5
#> 37    nec-ecx 2.37467820  0.4
#> 38    nec-ecx 2.43974293  0.3
#> 39    nec-ecx 2.50480766  0.2
#> 40    nec-ecx 2.56987239  0.2
#> 41    nec-ecx 2.63493712  0.2
#> 42    nec-ecx 2.70000184  0.1
#> 43    nec-ecx 2.76506657  0.1
#> 44    nec-ecx 2.83013130  0.1
#> 45    nec-ecx 2.89519603  0.1
#> 46    nec-ecx 2.96026075  0.1
#> 47    nec-ecx 3.02532548  0.1
#> 48    nec-ecx 3.09039021  0.1
#> 49    nec-ecx 3.15545494  0.1
#> 50    nec-ecx 3.22051966  0.1
#>