Posterior dispersion

Calculates posterior dispersion metric

dispersion(model, summary = FALSE, seed = 10)

Arguments

model
An object of class bayesnecfit whose distribution family is either poisson or binomial.

summary
Logical. Should summary stats be returned instead of full vector? Defaults to FALSE.

seed
Change seed for reproducible purposes.

Value

A numeric vector. If summary is FALSE, an n-long vector containing the dispersion metric, where n is the number of post warm-up posterior draws from the brmsfit object. If TRUE, then a data.frame containing the summary stats (mean, median, 95% highest density intervals) of the dispersion metric.

Details

This function calculates a dispersion metric which takes the ratio between the observed relative to simulated Pearson residuals sums of squares.

References

Zuur, A. F., Hilbe, J. M., & Ieno, E. N. (2013). A Beginner's Guide to GLM and GLMM with R: A Frequentist and Bayesian Perspective for Ecologists. Highland Statistics Limited.

Examples

# \donttest{
library(bayesnec)
data(nec_data)
nec_data$y <- as.integer(round(nec_data$y * 100))
nec4param <- bnec(y ~ crf(x, "nec4param"), data = nec_data, chains = 2)
#> Finding initial values which allow the response to be fitted using a nec4param model and a poisson distribution.
#> Compiling Stan program...
#> Start sampling
#> 
#> SAMPLING FOR MODEL '7872a07ff001f6c75a6939f1c914d8cc' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000106 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.06 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 10000 [  0%]  (Warmup)
#> Chain 1: Iteration: 1000 / 10000 [ 10%]  (Warmup)
#> Chain 1: Iteration: 2000 / 10000 [ 20%]  (Warmup)
#> Chain 1: Iteration: 3000 / 10000 [ 30%]  (Warmup)
#> Chain 1: Iteration: 4000 / 10000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 5000 / 10000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 6000 / 10000 [ 60%]  (Warmup)
#> Chain 1: Iteration: 7000 / 10000 [ 70%]  (Warmup)
#> Chain 1: Iteration: 8000 / 10000 [ 80%]  (Warmup)
#> Chain 1: Iteration: 8001 / 10000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 9000 / 10000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 10000 / 10000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 3.04812 seconds (Warm-up)
#> Chain 1:                0.897723 seconds (Sampling)
#> Chain 1:                3.94584 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL '7872a07ff001f6c75a6939f1c914d8cc' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 5.6e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.56 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 10000 [  0%]  (Warmup)
#> Chain 2: Iteration: 1000 / 10000 [ 10%]  (Warmup)
#> Chain 2: Iteration: 2000 / 10000 [ 20%]  (Warmup)
#> Chain 2: Iteration: 3000 / 10000 [ 30%]  (Warmup)
#> Chain 2: Iteration: 4000 / 10000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 5000 / 10000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 6000 / 10000 [ 60%]  (Warmup)
#> Chain 2: Iteration: 7000 / 10000 [ 70%]  (Warmup)
#> Chain 2: Iteration: 8000 / 10000 [ 80%]  (Warmup)
#> Chain 2: Iteration: 8001 / 10000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 9000 / 10000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 10000 / 10000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 3.08116 seconds (Warm-up)
#> Chain 2:                1.01753 seconds (Sampling)
#> Chain 2:                4.09868 seconds (Total)
#> Chain 2: 
#> Response variable modelled as a nec4param model using a poisson distribution.
#> Warning: Found 1 observations with a pareto_k > 0.7 in model 'nec4param'. It is recommended to set 'moment_match = TRUE' in order to perform moment matching for problematic observations. 
#> Warning: 
#> 1 (1.0%) p_waic estimates greater than 0.4. We recommend trying loo instead.
dispersion(nec4param, summary = TRUE)
#>    Estimate        Q2.5       Q97.5 
#>   5.0204670   0.8264864 136.8109087 
# }