fit_bayesnec

Fits a concentration(dose)-response model using brms

fit_bayesnec(
  data,
  x_var,
  y_var,
  trials_var = NA,
  family = NULL,
  priors,
  model = NA,
  inits,
  skip_check = FALSE,
  ...
)

Arguments

data
A data.frame containing the data to use for the model.

x_var
A character indicating the column heading containing the concentration (x) variable.

y_var
A character indicating the column heading containing the response (y) variable.

trials_var
The column heading indicating the column for the number of "trials" for binomial response data. If not supplied, the model may run but will not be the model you intended!

family
Either a character string, a function, or an object of class family defining the statistical distribution (family) to use for the y (response) data. See details.

priors
An object of class brmsprior which specifies user-desired prior distributions of model parameters. If missing, bnec will figure out a baseline prior for each parameter.

model
A character vector indicating the model(s) to fit. See Details for more information.

inits
Optional. Initialisation values. Must be a list of "n" names lists, where "n" corresponds to the number of chains, and names correspond to the parameter names of a given model.

skip_check
Should data check via check_data be avoided? Only relevant to function amend. Defaults to FALSE.

...
Further arguments to brm via fit_bayesnec.

Value

The fitted brms model, including an estimate of the NEC value and predicted posterior values. A posterior sample of the NEC is also available under nec_posterior

See also